feat: draw a proximity profile from raw CSV data of 2007
… pretty interesting !
