+++ app/src/main/java/com/illiouchine/jm/ui/composable/plot/ProximitySpider.kt
@@ -12,12 +12,21 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.illiouchine.jm.model.Ballot
+import com.illiouchine.jm.model.Grading
+import com.illiouchine.jm.model.Judgment
+import com.illiouchine.jm.model.Poll
+import com.illiouchine.jm.model.PollConfig
 import com.illiouchine.jm.service.ProximityAnalysis
 import com.illiouchine.jm.service.ProximityAnalyzer
 import com.illiouchine.jm.ui.composable.plot.lib.SpiderChart
 import com.illiouchine.jm.ui.composable.plot.utils.truncate
 import com.illiouchine.jm.ui.preview.PreviewDataFaker
 import com.illiouchine.jm.ui.theme.JmTheme
+import kotlinx.collections.immutable.toPersistentList
+import org.jetbrains.kotlinx.dataframe.DataFrame
+import org.jetbrains.kotlinx.dataframe.api.map
+import org.jetbrains.kotlinx.dataframe.io.readCsvStr
 import java.lang.Integer.min
 
 // Shows how close (in the collective hearts of the judges) pairs of proposals are.
@@ -44,7 +53,9 @@ fun ProximitySpider(
 
     SpiderChart(
         modifier = modifier,
-//        title = {},
+        title = {
+            Text("Proximity of candidates with ${analysis.proposals[selectedCategoryIndex]}")
+        },
         categories = filteredAnalysis.proposals.map { it.truncate(16, "…") },
         values = filteredAnalysis.proximities[selectedCategoryIndex].map { it.toFloat() },
         tickValues = listOf(-1f, 0f, 1f),
@@ -76,7 +87,7 @@ fun ProximitySpider(
 fun PreviewProximitySpider(modifier: Modifier = Modifier) {
     val poll = PreviewDataFaker.poll(
         amountOfBallots = 7,
-        amountOfProposals = 12,
+        amountOfProposals = 20,
     )
     val analyzer = ProximityAnalyzer()
     val analysis = analyzer.analyze(
@@ -86,6 +97,1228 @@ fun PreviewProximitySpider(modifier: Modifier = Modifier) {
         Column(modifier) {
             Text("\uD83E\uDD9D I am the glitch raccoon that looks for glitches:")
 //            PlotTitle("Proximity Spider\n${poll.pollConfig.subject}")
+            ProximitySpider(
+                modifier = Modifier.height(400.dp),
+                analysis = analysis,
+            )
+        }
+    }
+}
+
+@Preview(
+    name = "2007 Phone (Portrait)",
+    showSystemUi = false,
+    uiMode = Configuration.UI_MODE_NIGHT_YES,
+    fontScale = 1.0f,
+)
+@Preview(
+    name = "2007 Phone (Landscape)",
+    showSystemUi = false,
+    uiMode = Configuration.UI_MODE_NIGHT_YES,
+    device = "spec:width=511dp,height=891dp,orientation=landscape",
+    fontScale = 1.0f,
+)
+@Preview(
+    name = "2007 Square",
+    showSystemUi = false,
+    uiMode = Configuration.UI_MODE_NIGHT_YES,
+    device = "spec:width=500dp,height=450dp",
+)
+@Composable
+fun PreviewCsvProximitySpider(modifier: Modifier = Modifier) {
+    // Real data from 2007 (bureaux 1 & 6)
+    val csvString = """
+Besancenot,Buffet,Schivardi,Bayrou,Bové,Voynet,De Villiers,Royal,Nihous,Le Pen,Laguiller,Sarkozy
+5,4,0,0,4,3,0,2,0,0,0,0
+0,0,0,2,0,0,0,4,0,0,0,0
+0,0,0,0,0,0,0,1,0,1,0,0
+5,5,0,2,5,2,0,6,2,0,2,0
+2,0,1,6,0,5,4,4,1,4,3,0
+5,4,0,0,5,5,0,6,0,0,4,0
+4,4,1,5,5,6,4,4,1,1,5,0
+0,0,0,0,0,3,0,5,0,0,0,1
+6,0,0,3,4,0,1,4,0,1,0,1
+0,0,1,0,0,0,1,3,1,1,0,1
+1,4,1,4,2,5,1,5,1,1,0,1
+3,3,1,4,1,1,1,3,2,1,0,1
+3,1,1,4,3,1,2,1,1,1,1,1
+1,1,1,2,2,1,1,2,1,1,1,1
+1,3,1,4,1,2,1,2,1,1,1,1
+3,3,1,4,3,1,1,3,1,1,1,1
+2,2,1,5,1,2,1,3,1,1,1,1
+1,1,1,5,4,3,1,3,1,1,1,1
+2,1,1,6,3,3,3,3,1,1,1,1
+4,2,2,2,2,2,1,4,1,1,1,1
+1,1,1,3,1,4,1,4,1,1,1,1
+4,1,1,6,1,1,1,5,1,1,1,1
+1,1,1,4,1,2,1,5,1,1,1,1
+2,2,1,6,2,2,1,5,1,1,1,1
+1,1,1,4,2,2,2,5,1,1,1,1
+1,1,1,3,1,1,1,6,1,1,1,1
+2,2,1,6,2,2,1,6,1,1,1,1
+1,1,1,4,1,3,1,6,1,1,1,1
+1,3,1,5,1,3,1,6,1,1,1,1
+2,2,2,4,2,3,1,6,1,1,1,1
+2,3,1,6,4,4,2,6,1,1,1,1
+1,1,1,5,1,2,3,1,2,1,1,1
+5,3,2,5,2,3,4,1,2,1,1,1
+2,3,1,5,2,0,1,4,2,1,1,1
+2,1,1,4,2,2,1,4,2,1,1,1
+3,3,1,2,3,3,1,4,2,1,1,1
+3,3,3,5,3,6,1,5,2,1,1,1
+4,4,2,3,1,2,1,6,2,1,1,1
+5,5,2,6,2,3,1,6,2,1,1,1
+2,2,2,3,4,4,1,6,2,1,1,1
+2,3,3,4,2,5,1,6,2,1,1,1
+1,1,1,4,1,2,1,5,3,1,1,1
+1,1,1,5,2,2,2,1,4,1,1,1
+2,1,1,6,1,2,4,3,1,5,1,1
+2,2,1,2,1,2,1,2,0,1,2,1
+2,2,1,4,2,2,1,5,0,1,2,1
+1,1,1,6,1,2,1,6,0,1,2,1
+2,3,1,5,1,3,1,1,1,1,2,1
+5,3,1,5,1,5,1,1,1,1,2,1
+4,4,1,5,3,3,1,2,1,1,2,1
+5,6,4,0,6,5,1,2,1,1,2,1
+3,3,3,1,1,3,1,3,1,1,2,1
+2,0,2,5,4,5,2,3,1,1,2,1
+2,2,1,3,2,2,1,4,1,1,2,1
+3,2,2,4,4,3,1,4,1,1,2,1
+2,3,2,1,1,2,1,5,1,1,2,1
+5,3,2,4,2,2,1,5,1,1,2,1
+2,3,1,5,2,2,1,5,1,1,2,1
+2,3,1,5,2,2,1,5,1,1,2,1
+3,2,1,2,3,2,1,5,1,1,2,1
+3,3,2,0,4,3,1,5,1,1,2,1
+2,2,1,5,1,4,1,5,1,1,2,1
+2,3,1,3,2,4,1,5,1,1,2,1
+3,5,1,1,3,4,1,5,1,1,2,1
+2,2,1,3,2,4,1,6,1,1,2,1
+4,4,1,5,2,4,1,6,1,1,2,1
+4,3,4,5,4,1,1,1,2,1,2,1
+1,1,1,4,4,2,1,1,2,1,2,1
+2,2,2,3,2,2,2,1,2,1,2,1
+2,2,2,3,3,2,2,1,2,1,2,1
+5,3,4,5,3,3,2,1,2,1,2,1
+4,3,4,4,2,3,1,2,2,1,2,1
+3,2,2,6,2,3,1,5,2,1,2,1
+3,3,1,5,3,4,1,5,2,1,2,1
+2,2,2,3,4,2,2,5,2,1,2,1
+3,4,1,5,2,3,3,5,2,1,2,1
+4,5,2,6,3,4,1,6,2,1,2,1
+3,3,2,5,3,5,1,6,2,1,2,1
+3,2,2,5,2,3,1,4,3,1,2,1
+2,2,1,5,2,0,1,6,3,1,2,1
+5,4,1,2,5,2,1,2,1,1,3,1
+3,3,1,3,1,3,1,3,1,1,3,1
+5,6,4,2,5,4,1,3,1,1,3,1
+3,5,2,2,5,0,1,4,1,1,3,1
+5,2,1,3,1,1,1,4,1,1,3,1
+2,3,1,5,2,1,1,4,1,1,3,1
+4,4,1,5,2,1,1,4,1,1,3,1
+3,3,2,4,2,4,1,5,1,1,3,1
+3,4,1,4,3,4,1,5,1,1,3,1
+4,3,1,2,1,5,1,5,1,1,3,1
+5,2,1,2,4,5,1,5,1,1,3,1
+3,5,2,3,4,6,1,5,1,1,3,1
+3,3,1,5,2,2,2,5,1,1,3,1
+5,4,2,5,3,2,2,5,1,1,3,1
+1,2,1,5,1,2,1,6,1,1,3,1
+6,5,2,3,4,2,1,6,1,1,3,1
+4,4,1,2,5,2,1,6,1,1,3,1
+3,3,1,5,3,4,1,6,1,1,3,1
+2,4,1,1,1,5,1,6,1,1,3,1
+1,3,1,5,4,5,1,6,1,1,3,1
+3,3,3,5,4,5,1,6,1,1,3,1
+5,3,1,6,5,5,1,6,1,1,3,1
+5,1,2,5,3,4,1,3,2,1,3,1
+4,3,4,2,3,2,1,4,2,1,3,1
+4,5,2,1,4,4,1,6,2,1,3,1
+5,5,5,4,5,5,1,5,3,1,3,1
+5,5,5,4,3,3,1,6,3,1,3,1
+4,5,0,2,4,4,1,2,0,1,4,1
+5,4,1,6,4,2,1,6,0,1,4,1
+5,5,0,2,5,5,2,6,0,1,4,1
+5,4,1,3,2,2,1,3,1,1,4,1
+4,4,1,5,3,4,1,3,1,1,4,1
+5,4,1,6,5,4,1,3,1,1,4,1
+2,3,1,6,2,4,1,4,1,1,4,1
+6,6,2,2,5,4,1,4,1,1,4,1
+6,4,1,2,5,2,1,5,1,1,4,1
+4,4,3,4,2,5,1,5,1,1,4,1
+4,4,2,4,3,5,1,5,1,1,4,1
+4,6,2,4,4,5,1,5,1,1,4,1
+4,5,3,3,5,4,1,6,1,1,4,1
+4,4,1,1,1,5,1,6,1,1,4,1
+4,5,1,4,2,6,1,6,1,1,4,1
+3,5,4,4,4,6,1,6,1,1,4,1
+3,1,1,5,3,2,2,6,1,1,4,1
+6,5,4,5,5,6,4,3,2,1,4,1
+4,4,1,6,2,2,1,5,2,1,4,1
+5,4,2,3,5,3,1,3,3,1,4,1
+4,4,2,5,4,5,1,4,3,1,4,1
+4,3,4,0,4,5,3,3,4,1,4,1
+3,1,2,5,4,4,1,6,4,1,4,1
+6,4,1,5,2,2,1,5,2,2,4,1
+5,5,1,6,6,6,1,6,1,1,5,1
+5,6,2,1,4,4,1,5,2,1,5,1
+5,5,2,4,5,5,1,6,2,1,5,1
+6,4,2,3,3,4,1,6,3,1,5,1
+5,5,1,5,5,5,1,6,4,1,5,1
+6,2,2,4,5,6,2,2,3,6,5,1
+6,4,1,2,6,3,1,2,1,1,6,1
+6,5,1,4,6,1,1,5,1,1,6,1
+6,6,6,1,6,2,1,2,2,1,6,1
+6,4,2,1,2,3,1,2,2,1,6,1
+2,4,4,2,6,2,1,6,2,1,6,1
+5,5,1,4,3,5,2,6,2,1,6,1
+5,6,3,4,3,3,2,4,3,1,6,1
+4,3,1,4,3,4,1,5,1,1,0,2
+2,1,1,4,2,3,1,4,1,1,1,2
+1,2,1,6,1,1,1,5,1,1,1,2
+2,2,1,4,1,3,1,6,1,1,1,2
+2,2,1,4,1,3,1,2,1,1,2,2
+2,3,1,4,1,3,1,5,1,1,2,2
+1,2,1,3,2,2,1,4,2,1,2,2
+2,2,2,5,2,3,1,4,2,1,2,2
+0,1,1,5,3,2,1,4,1,1,3,2
+3,4,3,0,5,4,1,4,1,1,3,2
+2,2,1,4,2,3,1,6,1,1,3,2
+3,4,1,6,4,5,1,6,1,1,3,2
+4,5,3,4,5,6,1,3,2,1,3,2
+5,3,2,5,4,6,1,5,2,1,3,2
+4,4,2,2,3,3,2,6,2,1,3,2
+3,3,2,4,4,4,2,6,2,1,3,2
+4,3,2,5,3,5,3,6,2,1,3,2
+5,4,3,4,3,4,1,2,3,1,3,2
+5,3,3,4,2,4,1,5,3,1,3,2
+2,5,2,2,1,3,1,5,1,1,4,2
+5,3,1,5,2,1,1,2,2,1,4,2
+4,4,3,4,3,3,1,5,2,1,4,2
+4,4,1,4,4,4,1,6,2,1,4,2
+5,4,4,3,4,4,4,3,5,1,4,2
+6,5,3,3,4,3,1,4,0,1,5,2
+6,5,4,2,3,2,1,4,2,1,5,2
+5,5,4,4,5,3,3,5,5,3,5,2
+5,4,3,2,2,3,1,3,1,1,6,2
+6,5,2,3,5,5,1,4,2,1,6,2
+6,5,3,3,1,4,1,2,3,1,6,2
+4,4,0,5,4,4,1,5,0,0,0,3
+2,4,3,6,3,4,2,5,0,1,0,3
+1,2,1,4,1,2,1,0,1,1,1,3
+1,2,1,3,1,2,2,0,1,1,1,3
+1,1,1,5,1,1,3,1,1,1,1,3
+1,1,1,4,1,1,1,2,1,1,1,3
+1,2,1,6,1,4,1,2,1,1,1,3
+1,1,1,4,1,1,1,3,1,1,1,3
+2,3,1,5,2,4,1,3,1,1,1,3
+3,2,1,3,1,2,1,4,1,1,1,3
+1,1,1,5,1,2,1,4,1,1,1,3
+2,2,1,5,1,3,1,4,1,1,1,3
+2,1,1,4,2,4,1,4,1,1,1,3
+2,2,1,5,2,2,1,6,1,1,1,3
+0,0,1,3,1,3,1,6,1,1,1,3
+2,2,1,3,1,3,1,6,1,1,1,3
+1,3,1,6,3,6,1,6,1,1,1,3
+3,2,2,5,2,2,2,3,2,1,1,3
+1,2,2,3,2,2,1,4,2,1,1,3
+2,2,2,5,2,3,2,4,2,1,1,3
+1,1,5,5,1,2,2,6,2,1,1,3
+1,1,1,2,1,1,3,1,0,4,1,3
+2,1,1,5,2,2,2,2,1,1,2,3
+4,2,2,5,1,4,1,3,1,1,2,3
+1,2,1,5,1,1,2,3,1,1,2,3
+1,2,1,5,1,2,2,3,1,1,2,3
+2,2,1,5,2,3,1,4,1,1,2,3
+3,3,2,5,2,3,1,4,1,1,2,3
+3,1,1,3,3,3,1,4,1,1,2,3
+2,4,1,6,2,2,1,5,1,1,2,3
+2,2,2,2,5,2,2,2,2,1,2,3
+3,2,2,4,3,2,1,3,2,1,2,3
+1,0,1,5,1,3,1,3,2,1,2,3
+4,2,1,5,1,4,1,5,2,1,2,3
+4,2,2,5,2,5,1,5,2,1,2,3
+3,3,2,6,3,4,2,5,2,1,2,3
+2,2,2,5,0,4,1,6,2,1,2,3
+4,3,1,4,4,1,1,5,4,1,2,3
+2,2,2,5,1,3,1,5,1,1,3,3
+6,5,2,4,3,4,1,5,1,1,3,3
+3,3,1,4,3,4,1,6,1,1,3,3
+3,3,2,5,3,2,1,4,2,1,3,3
+2,2,2,3,3,3,1,4,2,1,3,3
+3,3,2,4,3,4,1,4,2,1,3,3
+3,4,2,5,2,4,1,6,2,1,3,3
+5,5,3,5,4,4,4,6,2,1,3,3
+3,3,3,3,2,3,3,3,3,1,3,3
+2,3,1,6,1,2,1,5,3,1,3,3
+5,6,2,6,5,5,1,5,3,1,3,3
+1,3,2,5,1,3,1,6,3,1,3,3
+3,3,3,4,4,2,1,6,5,1,3,3
+4,5,2,3,2,3,1,4,2,1,4,3
+3,4,1,5,3,4,1,6,2,1,4,3
+4,4,0,3,2,3,2,4,0,3,4,3
+3,2,1,3,1,1,2,2,1,5,4,3
+5,4,3,4,4,4,0,5,3,5,4,3
+1,1,2,2,1,2,3,1,0,6,4,3
+5,6,4,4,5,4,3,4,4,3,5,3
+5,4,3,3,3,2,1,1,3,4,5,3
+1,1,1,4,1,1,1,1,1,1,1,4
+1,1,1,3,4,1,2,1,1,1,1,4
+1,1,1,4,1,2,0,2,1,1,1,4
+1,1,1,2,1,1,1,2,1,1,1,4
+1,2,3,5,1,1,1,2,1,1,1,4
+1,1,1,5,1,2,1,3,1,1,1,4
+2,2,1,3,2,2,1,3,1,1,1,4
+1,1,1,5,3,2,1,3,1,1,1,4
+1,1,1,3,1,3,1,3,1,1,1,4
+1,1,1,4,1,3,1,3,1,1,1,4
+1,1,1,2,1,2,2,3,1,1,1,4
+1,2,1,5,2,2,2,3,1,1,1,4
+2,2,1,5,2,3,2,3,1,1,1,4
+1,2,1,4,1,1,3,3,1,1,1,4
+1,1,1,5,1,2,3,3,1,1,1,4
+1,1,1,4,1,1,1,4,1,1,1,4
+1,1,1,5,1,1,1,4,1,1,1,4
+3,3,1,3,1,2,1,4,1,1,1,4
+1,1,1,5,1,2,1,4,1,1,1,4
+1,1,1,5,2,3,1,4,1,1,1,4
+3,3,1,5,2,4,1,4,1,1,1,4
+2,0,1,5,1,2,2,4,1,1,1,4
+3,3,3,6,1,2,2,4,1,1,1,4
+2,2,1,1,2,2,2,4,1,1,1,4
+2,2,1,5,2,2,2,4,1,1,1,4
+2,1,1,4,2,2,1,5,1,1,1,4
+3,3,1,4,1,3,1,5,1,1,1,4
+2,1,1,4,3,3,1,5,1,1,1,4
+3,2,3,6,4,5,1,5,1,1,1,4
+3,3,1,5,1,3,5,5,1,1,1,4
+1,1,1,1,1,4,1,6,1,1,1,4
+2,2,1,5,2,4,1,6,1,1,1,4
+1,2,1,1,2,2,2,6,1,1,1,4
+1,2,1,5,3,1,1,2,2,1,1,4
+2,1,2,4,1,1,3,3,2,1,1,4
+1,1,2,3,2,2,3,3,2,1,1,4
+1,2,2,5,3,2,1,4,2,1,1,4
+2,1,1,6,1,3,1,5,2,1,1,4
+1,2,1,5,1,2,1,0,4,1,1,4
+1,1,1,4,1,2,2,1,4,1,1,4
+1,1,3,3,3,3,1,1,1,2,1,4
+1,1,1,3,2,3,2,2,1,2,1,4
+1,1,1,3,1,1,1,3,3,2,1,4
+2,3,2,4,1,5,1,3,2,3,1,4
+1,2,2,4,1,2,2,2,1,1,2,4
+2,2,1,4,2,2,0,3,1,1,2,4
+2,1,1,5,2,2,1,3,1,1,2,4
+2,2,2,6,2,2,2,3,1,1,2,4
+1,2,1,6,1,1,1,4,1,1,2,4
+2,2,1,4,2,3,1,4,1,1,2,4
+2,3,1,5,3,3,1,4,1,1,2,4
+2,2,1,5,3,4,2,4,1,1,2,4
+2,0,1,6,3,3,3,4,1,1,2,4
+3,1,1,6,0,1,1,5,1,1,2,4
+3,2,1,5,3,2,1,5,1,1,2,4
+1,2,1,4,4,4,1,5,1,1,2,4
+2,2,1,4,4,4,2,5,1,1,2,4
+2,3,1,5,4,3,3,5,1,1,2,4
+3,3,1,4,2,3,1,6,1,1,2,4
+4,4,2,5,3,3,1,6,1,1,2,4
+2,2,2,5,3,5,2,6,1,1,2,4
+2,3,2,6,2,2,2,2,2,1,2,4
+1,2,2,3,1,2,1,3,2,1,2,4
+2,2,2,3,2,2,1,3,2,1,2,4
+2,1,1,4,1,3,4,3,2,1,2,4
+2,2,2,4,2,2,1,4,2,1,2,4
+1,3,1,4,3,3,1,6,2,1,2,4
+3,3,3,5,5,5,1,5,3,1,2,4
+2,2,1,6,3,3,1,3,4,1,2,4
+4,4,2,6,2,2,1,4,4,1,2,4
+4,3,2,5,2,1,3,4,1,2,2,4
+4,4,1,5,3,4,0,2,2,2,2,4
+2,2,2,5,0,0,3,3,2,2,2,4
+2,2,1,3,2,2,3,4,2,3,2,4
+4,4,2,5,3,3,2,5,2,3,2,4
+4,3,1,4,1,2,3,3,1,4,2,4
+2,2,1,3,1,3,1,4,1,1,3,4
+2,3,2,6,2,3,1,5,1,1,3,4
+3,2,2,5,3,4,1,6,2,1,3,4
+4,2,2,6,2,5,5,6,2,1,3,4
+4,3,1,4,3,4,1,3,3,1,3,4
+3,4,1,6,1,1,1,4,3,1,3,4
+3,3,1,5,3,4,1,5,3,1,3,4
+4,4,0,5,4,5,4,6,0,2,3,4
+1,1,1,3,1,3,1,2,2,3,3,4
+3,4,1,4,4,3,3,4,2,4,3,4
+3,2,1,4,1,2,5,3,4,6,3,4
+2,4,1,6,4,5,1,6,1,1,4,4
+2,4,4,6,4,5,4,4,2,1,4,4
+4,2,1,4,2,2,1,5,2,1,5,4
+5,1,1,5,4,3,2,5,1,2,5,4
+0,0,0,0,0,0,0,0,0,0,0,5
+0,0,0,4,0,0,0,3,0,0,0,5
+0,0,0,5,0,0,1,3,0,1,0,5
+0,1,0,3,1,0,0,4,0,1,0,5
+1,3,0,4,0,0,4,2,0,3,0,5
+0,0,0,3,2,2,0,3,0,3,0,5
+1,1,1,6,1,2,2,2,0,1,1,5
+1,1,1,6,3,1,1,1,1,1,1,5
+2,1,1,3,1,1,2,1,1,1,1,5
+1,1,1,1,1,1,4,1,1,1,1,5
+1,1,1,5,1,1,1,2,1,1,1,5
+1,1,1,6,1,2,1,2,1,1,1,5
+1,1,1,5,1,1,4,2,1,1,1,5
+1,2,1,4,1,1,1,3,1,1,1,5
+1,1,1,5,1,1,1,3,1,1,1,5
+1,1,1,6,1,2,1,3,1,1,1,5
+1,1,1,6,1,2,2,3,1,1,1,5
+1,2,1,4,2,3,3,3,1,1,1,5
+1,1,1,6,5,1,1,4,1,1,1,5
+3,3,2,5,2,4,1,4,1,1,1,5
+3,2,1,5,1,2,3,4,1,1,1,5
+1,1,1,6,3,1,0,5,1,1,1,5
+2,4,1,6,1,2,1,5,1,1,1,5
+4,1,1,5,5,5,1,5,1,1,1,5
+1,2,1,5,1,1,2,5,1,1,1,5
+2,1,2,4,1,3,2,1,2,1,1,5
+1,1,1,2,1,1,1,2,2,1,1,5
+1,1,1,5,2,2,2,3,2,1,1,5
+2,2,1,6,1,3,3,3,2,1,1,5
+1,1,1,3,2,2,3,4,2,1,1,5
+1,2,1,6,1,3,4,4,2,1,1,5
+1,1,1,6,1,2,3,5,2,1,1,5
+1,2,1,4,1,2,3,3,3,1,1,5
+1,1,2,3,1,2,6,2,1,2,1,5
+3,1,1,3,1,4,2,3,1,2,1,5
+3,3,2,4,3,3,4,4,1,2,1,5
+1,1,1,4,1,0,6,1,1,3,1,5
+1,2,1,4,2,3,2,2,1,3,1,5
+1,3,2,4,1,3,4,3,1,3,1,5
+1,1,1,5,1,1,1,4,1,3,1,5
+1,1,1,4,1,1,5,1,2,3,1,5
+1,2,1,4,1,2,5,3,2,3,1,5
+2,1,1,3,1,1,6,3,2,3,1,5
+1,1,1,4,1,2,3,3,3,3,1,5
+1,1,1,3,1,1,3,2,1,4,1,5
+3,1,1,4,1,2,4,2,1,4,1,5
+1,1,1,4,1,2,4,3,4,4,1,5
+1,1,1,1,1,1,6,1,5,5,1,5
+1,1,1,4,1,4,5,5,5,5,1,5
+1,1,1,1,1,2,5,1,1,6,1,5
+1,0,1,0,1,2,2,1,5,6,1,5
+2,2,2,6,1,2,2,3,0,1,2,5
+4,2,1,5,2,2,3,5,0,1,2,5
+1,2,0,3,4,6,1,0,1,1,2,5
+2,2,1,5,1,2,1,3,1,1,2,5
+3,4,3,4,1,2,2,3,1,1,2,5
+3,3,2,4,2,2,1,4,1,1,2,5
+3,4,1,5,2,4,1,5,1,1,2,5
+2,2,1,5,1,3,1,6,1,1,2,5
+2,2,1,5,1,2,2,6,1,1,2,5
+2,1,1,6,1,2,3,2,2,1,2,5
+4,3,2,4,2,4,3,4,2,1,2,5
+1,1,1,5,2,3,4,4,2,1,2,5
+2,1,1,6,2,3,1,5,2,1,2,5
+3,2,1,6,2,4,1,5,2,1,2,5
+2,3,2,3,3,4,1,5,2,1,2,5
+2,2,2,4,2,2,3,5,2,1,2,5
+1,3,3,4,2,2,5,5,2,1,2,5
+3,3,4,6,2,3,1,6,2,1,2,5
+3,3,2,5,1,3,3,1,3,1,2,5
+2,1,1,5,1,3,1,3,3,1,2,5
+1,2,2,5,2,3,5,3,3,1,2,5
+3,2,3,5,4,3,3,5,3,1,2,5
+3,2,1,3,2,2,1,3,1,2,2,5
+1,2,1,5,1,2,2,4,1,2,2,5
+1,1,1,3,1,1,1,1,1,3,2,5
+2,2,1,3,1,2,3,3,1,3,2,5
+1,2,1,6,2,3,2,4,1,3,2,5
+3,2,1,5,2,4,4,5,1,3,2,5
+3,3,2,6,2,4,0,5,2,3,2,5
+2,3,1,5,1,2,4,4,3,3,2,5
+1,2,1,4,1,1,6,2,1,4,2,5
+3,4,1,6,1,4,2,3,1,4,2,5
+2,3,1,4,2,2,2,4,1,4,2,5
+4,3,1,5,2,4,2,2,1,1,3,5
+1,2,1,5,2,2,1,3,1,1,3,5
+3,3,2,4,2,3,5,4,1,1,3,5
+1,1,1,5,3,1,1,5,1,1,3,5
+2,3,3,4,2,3,5,2,2,1,3,5
+4,2,2,2,4,2,2,3,2,1,3,5
+2,3,1,5,4,4,2,4,2,1,3,5
+3,3,2,4,4,3,3,5,2,1,3,5
+3,4,3,6,3,2,2,4,3,1,3,5
+2,3,3,5,1,4,3,6,3,1,3,5
+3,3,2,0,2,3,0,4,4,1,3,5
+3,3,3,4,3,3,3,3,3,3,3,5
+2,3,2,5,1,3,2,5,2,4,3,5
+6,3,2,5,2,2,1,6,1,1,4,5
+4,1,1,6,1,5,1,5,2,1,4,5
+6,4,1,5,3,4,1,5,3,1,4,5
+5,4,3,5,4,6,3,5,3,1,4,5
+2,1,2,3,1,1,3,3,2,4,4,5
+5,5,2,4,5,5,4,4,2,4,4,5
+5,5,1,4,4,4,5,5,1,6,4,5
+4,5,2,5,4,5,4,6,2,1,5,5
+4,2,1,2,1,5,2,3,1,3,5,5
+3,3,2,4,4,5,6,3,5,5,5,5
+5,2,2,4,2,2,1,4,2,1,6,5
+0,0,0,0,0,0,0,0,0,0,0,6
+0,0,0,3,0,0,0,0,0,0,0,6
+2,3,3,4,2,0,0,2,3,0,0,6
+2,1,3,2,0,1,3,0,0,1,0,6
+3,0,0,0,0,0,5,0,0,4,0,6
+1,2,1,5,2,3,3,3,2,4,0,6
+4,4,2,4,1,2,4,1,4,4,0,6
+1,1,1,1,1,1,1,1,1,1,1,6
+1,1,1,2,1,1,1,1,1,1,1,6
+1,1,1,4,1,1,1,1,1,1,1,6
+1,1,1,1,1,1,2,1,1,1,1,6
+1,1,1,2,1,2,2,1,1,1,1,6
+1,1,1,5,1,2,6,1,1,1,1,6
+1,1,1,3,1,1,1,2,1,1,1,6
+1,1,1,4,1,2,1,2,1,1,1,6
+1,1,1,6,1,1,2,2,1,1,1,6
+2,1,1,2,1,2,3,2,1,1,1,6
+1,1,1,5,1,4,6,2,1,1,1,6
+1,1,1,3,1,1,1,3,1,1,1,6
+1,1,1,5,1,3,1,3,1,1,1,6
+1,2,1,5,2,4,2,3,1,1,1,6
+1,1,2,3,1,1,3,3,1,1,1,6
+1,1,1,5,1,1,3,3,1,1,1,6
+1,1,2,4,3,1,1,4,1,1,1,6
+1,1,1,6,1,1,1,5,1,1,1,6
+2,2,1,6,1,1,1,5,1,1,1,6
+4,4,1,5,1,2,3,5,1,1,1,6
+1,1,1,5,4,4,1,6,1,1,1,6
+1,1,1,5,1,2,3,1,2,1,1,6
+1,1,1,4,1,2,1,2,2,1,1,6
+2,2,1,2,1,1,2,2,2,1,1,6
+2,2,1,3,1,3,4,2,2,1,1,6
+2,2,2,3,1,1,2,3,2,1,1,6
+1,3,1,5,1,3,2,3,2,1,1,6
+2,1,1,5,1,3,1,4,2,1,1,6
+1,2,2,5,2,4,2,4,2,1,1,6
+3,2,1,5,2,0,3,4,2,1,1,6
+3,1,2,0,1,4,4,4,2,1,1,6
+1,1,2,5,2,3,4,5,2,1,1,6
+1,1,2,6,1,1,1,2,3,1,1,6
+1,1,1,5,1,3,3,3,3,1,1,6
+3,3,2,5,1,4,1,4,3,1,1,6
+1,2,1,5,1,2,4,2,4,1,1,6
+1,1,1,1,1,1,1,1,1,2,1,6
+1,1,1,2,1,1,2,1,1,2,1,6
+1,1,1,5,1,2,2,3,1,2,1,6
+2,2,1,4,2,2,2,3,1,2,1,6
+1,1,1,4,1,1,4,3,1,2,1,6
+4,3,0,4,1,3,3,4,1,2,1,6
+0,1,1,4,1,1,3,1,2,2,1,6
+1,2,2,3,2,3,4,2,2,2,1,6
+2,1,1,3,1,1,5,2,2,2,1,6
+1,2,1,4,1,2,2,3,2,2,1,6
+5,1,1,6,2,1,1,1,1,3,1,6
+1,1,1,1,1,2,5,1,1,3,1,6
+1,1,1,1,1,1,2,2,1,3,1,6
+1,2,1,3,1,2,4,2,1,3,1,6
+1,1,1,4,1,1,4,5,1,3,1,6
+1,1,1,4,1,1,3,4,2,3,1,6
+1,1,1,4,1,1,1,4,4,3,1,6
+1,1,1,5,0,2,3,4,4,3,1,6
+1,1,4,4,1,2,3,4,4,3,1,6
+1,1,1,5,1,1,4,1,1,4,1,6
+1,1,1,2,1,1,5,1,1,4,1,6
+1,1,1,2,1,1,5,2,1,4,1,6
+1,2,2,5,1,2,4,4,1,4,1,6
+1,1,1,4,1,1,4,3,2,4,1,6
+1,1,2,2,2,2,5,3,2,4,1,6
+1,1,1,1,1,1,1,1,1,5,1,6
+1,1,1,2,1,1,5,1,1,5,1,6
+1,1,1,2,1,1,5,1,2,5,1,6
+2,2,1,5,2,2,4,2,2,5,1,6
+1,1,1,0,1,1,5,1,4,5,1,6
+1,1,2,4,1,1,6,5,4,5,1,6
+1,1,1,3,1,2,6,1,3,6,1,6
+1,1,1,4,1,1,5,3,4,6,1,6
+2,1,1,5,2,1,1,2,1,1,2,6
+1,1,1,2,2,2,1,2,1,1,2,6
+2,1,1,4,1,2,3,2,1,1,2,6
+1,1,1,3,1,1,1,3,1,1,2,6
+1,1,1,3,1,2,1,4,1,1,2,6
+1,1,2,5,1,3,1,2,2,1,2,6
+2,1,2,4,1,2,4,2,2,1,2,6
+2,3,2,6,2,3,4,3,2,1,2,6
+3,3,2,4,2,5,4,3,2,1,2,6
+1,3,1,4,1,2,2,4,2,1,2,6
+1,1,1,5,1,3,2,1,3,1,2,6
+3,3,1,6,2,3,2,6,3,1,2,6
+2,2,2,1,2,2,2,1,2,2,2,6
+4,3,3,5,3,3,4,2,2,2,2,6
+3,3,1,4,1,2,2,6,2,2,2,6
+1,2,1,1,1,3,5,1,1,3,2,6
+2,2,2,3,2,2,2,2,2,3,2,6
+4,4,3,3,4,4,3,3,2,3,2,6
+1,3,2,4,3,2,4,4,2,3,2,6
+1,2,1,4,1,2,1,4,1,4,2,6
+1,1,2,5,1,1,3,1,2,4,2,6
+2,2,2,4,3,2,3,4,2,4,2,6
+3,3,4,5,1,3,4,1,3,4,2,6
+2,2,2,4,2,3,4,4,5,4,2,6
+2,2,2,5,2,2,5,4,2,5,2,6
+1,0,1,3,1,4,4,2,3,6,2,6
+1,1,1,5,2,3,5,1,1,1,3,6
+1,1,1,2,2,2,1,4,1,1,3,6
+3,4,3,5,2,2,2,4,2,1,3,6
+1,1,1,5,1,0,1,4,3,1,3,6
+4,3,2,4,3,4,4,1,2,3,3,6
+3,2,2,3,2,2,3,5,2,3,3,6
+4,1,1,6,4,1,1,2,1,4,3,6
+2,2,3,5,3,3,4,5,1,4,3,6
+2,2,2,2,3,4,4,1,2,4,3,6
+3,2,2,3,0,3,4,2,2,4,3,6
+2,1,3,1,1,1,5,2,4,5,3,6
+4,3,1,3,1,3,1,3,4,5,3,6
+1,2,1,5,2,2,1,1,1,1,4,6
+4,5,1,6,5,5,2,6,1,4,4,6
+5,3,1,6,1,2,3,3,2,1,5,6
+4,1,1,6,4,5,5,4,4,1,5,6
+5,4,1,5,2,3,3,4,1,3,5,6
+4,3,3,5,5,3,6,4,3,3,5,6
+2,5,2,4,1,5,4,5,1,4,5,6
+4,3,2,5,4,3,3,3,2,4,5,6
+4,4,2,1,1,4,5,4,3,4,5,6
+3,4,2,5,3,4,4,1,1,5,5,6
+5,5,4,5,4,5,5,5,4,6,5,6
+1,1,1,3,1,1,6,4,5,6,5,6
+6,6,6,6,6,6,6,6,6,6,6,6
+2,0,0,0,0,0,0,0,0,0,0,0
+0,0,0,5,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,6,0,0,0,0
+0,0,0,0,0,0,0,6,0,0,0,0
+5,5,0,0,0,0,0,6,0,0,0,0
+2,3,2,5,1,3,2,5,4,2,0,0
+1,1,0,0,0,0,1,0,0,1,1,0
+3,1,1,4,1,1,1,2,1,1,1,0
+3,3,2,6,5,4,1,4,3,1,3,0
+3,4,3,6,2,4,0,4,0,0,6,0
+3,0,0,0,2,1,0,4,0,1,0,1
+4,6,2,2,2,2,1,5,1,1,0,1
+5,5,2,2,1,2,1,2,1,4,0,1
+3,2,0,2,2,6,0,6,0,1,1,1
+1,1,1,1,1,1,1,1,1,1,1,1
+1,1,1,1,1,1,1,1,1,1,1,1
+1,1,1,1,1,1,1,1,1,1,1,1
+4,6,3,2,3,1,1,1,1,1,1,1
+4,4,3,2,4,1,1,1,1,1,1,1
+4,4,1,4,3,4,1,1,1,1,1,1
+1,2,3,4,4,2,1,2,1,1,1,1
+4,5,1,2,4,2,1,3,1,1,1,1
+2,1,1,2,2,3,1,3,1,1,1,1
+2,1,1,6,3,4,1,3,1,1,1,1
+1,1,1,6,1,1,1,4,1,1,1,1
+5,4,1,5,2,1,1,4,1,1,1,1
+5,2,1,5,1,2,1,4,1,1,1,1
+3,1,1,5,2,2,1,4,1,1,1,1
+1,1,1,5,1,3,1,4,1,1,1,1
+1,1,2,3,2,3,1,4,1,1,1,1
+1,1,1,6,3,4,1,4,1,1,1,1
+6,5,3,2,1,1,1,5,1,1,1,1
+1,1,1,3,1,2,1,5,1,1,1,1
+1,2,1,5,2,2,1,5,1,1,1,1
+2,6,1,3,2,4,1,5,1,1,1,1
+1,1,1,6,1,5,1,5,1,1,1,1
+1,1,1,6,1,5,1,5,1,1,1,1
+1,1,1,1,1,1,1,6,1,1,1,1
+1,1,1,3,1,1,1,6,1,1,1,1
+4,1,1,5,1,1,1,6,1,1,1,1
+1,1,1,5,2,1,1,6,1,1,1,1
+3,5,2,4,2,3,1,6,1,1,1,1
+5,3,1,3,3,3,1,6,1,1,1,1
+3,2,1,1,1,4,1,6,1,1,1,1
+4,6,3,3,3,6,1,6,1,1,1,1
+2,2,1,4,1,2,2,6,1,1,1,1
+2,2,1,5,2,2,1,1,2,1,1,1
+3,3,3,5,2,2,2,3,2,1,1,1
+2,2,1,5,2,3,1,5,2,1,1,1
+3,2,3,3,1,2,3,5,2,1,1,1
+4,4,3,4,5,2,1,6,2,1,1,1
+3,3,2,6,2,5,1,6,2,1,1,1
+4,2,1,3,4,4,1,6,3,1,1,1
+2,2,2,4,1,3,0,6,4,1,1,1
+1,3,1,4,2,3,1,6,4,1,1,1
+5,3,2,6,1,1,1,2,6,1,1,1
+1,1,1,1,1,1,1,1,1,6,1,1
+4,5,0,3,5,3,1,0,0,1,2,1
+2,5,2,3,2,4,1,6,0,1,2,1
+3,1,5,3,2,2,1,1,1,1,2,1
+4,6,1,1,4,3,1,3,1,1,2,1
+4,3,2,2,2,4,1,3,1,1,2,1
+2,2,1,2,2,3,1,4,1,1,2,1
+2,2,1,4,2,3,1,4,1,1,2,1
+3,2,1,4,2,3,1,4,1,1,2,1
+3,2,1,4,2,4,1,4,1,1,2,1
+2,2,2,3,3,4,1,4,1,1,2,1
+5,5,2,2,3,6,1,4,1,1,2,1
+1,2,1,4,1,1,1,5,1,1,2,1
+2,3,1,4,1,2,1,5,1,1,2,1
+2,3,2,4,2,2,1,5,1,1,2,1
+2,4,1,4,2,4,1,5,1,1,2,1
+3,4,1,5,2,4,1,5,1,1,2,1
+6,4,3,1,6,5,1,5,1,1,2,1
+2,2,4,4,2,6,1,5,1,1,2,1
+2,2,1,5,2,1,1,6,1,1,2,1
+2,3,1,2,2,2,1,6,1,1,2,1
+2,2,2,3,2,2,1,6,1,1,2,1
+2,3,1,4,2,3,1,6,1,1,2,1
+5,4,2,1,5,3,1,6,1,1,2,1
+2,2,1,5,3,4,1,6,1,1,2,1
+2,3,1,5,2,5,2,6,1,1,2,1
+2,2,2,2,2,2,1,2,2,1,2,1
+2,2,2,5,3,2,1,4,2,1,2,1
+2,3,2,6,2,3,1,4,2,1,2,1
+3,2,2,6,3,3,1,4,2,1,2,1
+2,2,1,4,1,2,1,5,2,1,2,1
+2,2,1,6,2,2,1,5,2,1,2,1
+3,4,2,5,4,3,1,5,2,1,2,1
+6,4,2,4,5,3,1,5,2,1,2,1
+2,3,2,4,2,4,1,5,2,1,2,1
+4,4,2,3,4,4,1,5,2,1,2,1
+2,3,0,5,1,5,1,5,2,1,2,1
+4,4,2,4,2,5,1,5,2,1,2,1
+3,2,2,4,5,6,0,6,2,1,2,1
+6,6,2,2,2,2,1,6,2,1,2,1
+2,2,1,2,1,3,1,6,2,1,2,1
+3,3,3,2,3,3,1,6,2,1,2,1
+2,2,2,3,3,3,1,6,2,1,2,1
+3,5,2,3,4,5,1,6,2,1,2,1
+3,3,2,5,3,4,2,6,2,1,2,1
+6,3,2,6,4,5,3,6,2,1,2,1
+2,2,2,4,1,3,2,2,3,1,2,1
+4,5,3,6,4,5,2,4,3,1,2,1
+5,4,1,5,1,1,1,5,3,1,2,1
+3,4,1,2,3,2,1,5,4,1,2,1
+1,5,3,4,2,3,4,3,2,5,2,1
+2,3,1,1,6,2,1,1,1,1,3,1
+4,4,2,5,4,1,1,2,1,1,3,1
+6,5,3,2,0,4,1,2,1,1,3,1
+3,3,2,4,4,4,1,2,1,1,3,1
+5,2,2,1,4,6,1,2,1,1,3,1
+5,3,1,5,5,6,1,2,1,1,3,1
+5,5,1,1,2,3,1,3,1,1,3,1
+3,4,1,6,3,3,1,4,1,1,3,1
+3,3,2,4,3,4,1,4,1,1,3,1
+4,4,2,2,4,5,1,4,1,1,3,1
+3,4,3,4,3,3,1,5,1,1,3,1
+4,4,1,5,3,5,1,5,1,1,3,1
+3,6,1,3,1,6,1,5,1,1,3,1
+4,5,2,4,2,3,2,5,1,1,3,1
+4,3,2,5,4,6,2,5,1,1,3,1
+3,3,3,3,2,1,1,6,1,1,3,1
+4,3,1,3,3,3,1,6,1,1,3,1
+4,5,2,5,4,3,1,6,1,1,3,1
+4,4,0,4,4,4,1,6,1,1,3,1
+5,4,1,4,4,5,1,6,1,1,3,1
+3,3,1,5,3,3,2,6,1,1,3,1
+3,4,2,6,3,3,2,6,1,1,3,1
+5,4,3,5,3,3,1,3,2,1,3,1
+4,3,2,1,3,2,1,4,2,1,3,1
+4,2,1,4,3,3,1,5,2,1,3,1
+2,4,1,3,2,4,1,5,2,1,3,1
+5,3,1,6,3,5,1,5,2,1,3,1
+4,3,2,5,1,5,1,6,2,1,3,1
+4,4,1,5,2,5,1,6,2,1,3,1
+4,3,2,5,3,3,2,6,2,1,3,1
+3,3,1,6,1,5,1,2,3,1,3,1
+4,6,3,5,5,5,1,6,3,1,3,1
+2,4,4,4,5,5,1,6,4,1,3,1
+5,0,0,5,0,0,0,6,0,1,4,1
+2,2,1,4,2,2,1,5,1,1,4,1
+4,3,1,5,3,3,1,5,1,1,4,1
+6,5,4,3,4,3,1,5,1,1,4,1
+6,5,2,3,4,5,1,5,1,1,4,1
+6,6,2,1,5,5,1,5,1,1,4,1
+5,6,4,4,4,6,1,5,1,1,4,1
+3,5,5,3,2,2,1,6,1,1,4,1
+2,5,1,5,1,4,1,6,1,1,4,1
+4,5,2,4,2,4,1,6,1,1,4,1
+4,4,2,3,3,4,1,6,1,1,4,1
+5,4,2,3,2,6,1,3,2,1,4,1
+1,2,1,5,3,2,5,3,2,1,4,1
+6,6,4,3,4,5,1,5,2,1,4,1
+6,4,2,3,4,3,1,6,2,1,4,1
+5,5,1,3,4,4,1,6,2,1,4,1
+5,5,2,2,5,5,1,6,2,1,4,1
+4,4,1,3,3,6,1,6,2,1,4,1
+5,5,2,4,4,4,2,6,2,1,4,1
+4,5,3,6,6,6,2,6,2,1,4,1
+5,4,3,5,5,5,1,4,3,1,4,1
+5,5,3,5,2,3,1,6,3,1,4,1
+6,5,3,2,1,4,1,6,3,1,4,1
+1,4,3,6,4,5,4,6,3,1,4,1
+2,4,2,5,3,3,1,5,5,1,4,1
+4,6,4,1,4,4,1,4,0,1,5,1
+6,0,0,5,0,0,0,6,0,1,5,1
+6,5,2,2,4,3,1,2,1,1,5,1
+4,4,4,4,4,4,1,5,1,1,5,1
+5,6,1,5,5,5,1,5,1,1,5,1
+3,3,1,5,2,2,1,6,1,1,5,1
+5,5,1,2,3,4,1,6,1,1,5,1
+5,5,0,4,6,4,1,3,2,1,5,1
+5,5,3,3,6,6,1,5,2,1,5,1
+5,3,1,5,4,2,1,6,2,1,5,1
+6,4,3,6,5,4,1,6,2,1,5,1
+5,4,3,6,0,5,2,5,3,1,5,1
+4,4,3,6,5,5,5,5,4,1,5,1
+5,6,4,6,5,4,1,6,1,2,5,1
+2,3,1,4,2,3,3,2,1,5,5,1
+0,6,0,5,0,0,0,6,0,1,6,1
+6,4,1,1,1,1,1,1,1,1,6,1
+5,3,2,5,5,3,1,5,2,1,6,1
+6,5,3,4,5,5,1,5,2,1,6,1
+6,4,3,1,5,5,1,4,4,1,6,1
+5,6,4,6,5,5,1,5,5,5,6,1
+0,0,0,0,0,0,0,2,0,1,0,2
+1,4,2,3,4,1,0,6,0,1,1,2
+1,1,1,5,1,3,1,2,1,1,1,2
+4,5,1,4,4,4,1,2,1,1,1,2
+2,3,1,4,2,1,1,3,1,1,1,2
+2,2,1,5,2,2,1,4,1,1,1,2
+1,4,1,5,2,3,1,5,1,1,1,2
+4,4,1,3,2,4,1,5,1,1,1,2
+1,2,1,4,4,4,1,5,1,1,1,2
+1,2,1,5,1,5,1,5,1,1,1,2
+1,1,1,2,3,3,6,6,1,1,1,2
+2,2,2,2,2,2,1,3,2,1,1,2
+1,1,2,5,1,1,1,3,3,1,1,2
+2,2,2,5,3,3,1,2,4,1,1,2
+2,1,3,2,1,1,1,2,2,2,1,2
+2,2,1,4,2,2,1,3,1,1,2,2
+3,2,1,5,2,2,1,5,1,1,2,2
+3,3,2,3,2,3,1,5,1,1,2,2
+3,2,1,6,2,3,1,5,1,1,2,2
+2,2,1,4,2,4,1,5,1,1,2,2
+1,1,1,5,1,3,1,6,1,1,2,2
+3,3,2,4,3,5,1,6,1,1,2,2
+2,2,2,2,2,2,1,2,2,1,2,2
+2,2,2,5,2,2,1,4,2,1,2,2
+2,2,1,2,2,3,1,4,2,1,2,2
+2,3,1,5,1,2,1,5,2,1,2,2
+3,2,2,4,2,3,1,5,2,1,2,2
+3,3,3,3,2,2,1,5,3,2,2,2
+4,4,2,6,4,2,2,6,4,3,2,2
+2,2,1,3,2,2,1,2,1,1,3,2
+4,3,1,5,4,4,1,3,1,1,3,2
+1,3,1,4,1,3,1,5,1,1,3,2
+5,6,2,3,1,6,1,5,1,1,3,2
+3,3,2,6,1,4,1,3,2,1,3,2
+5,4,2,4,5,6,1,6,3,1,3,2
+3,3,2,6,1,1,1,3,4,1,3,2
+4,5,1,3,3,4,1,5,1,1,4,2
+3,4,2,5,4,4,3,4,2,1,4,2
+5,4,3,3,3,4,1,5,2,1,4,2
+4,5,2,4,4,4,1,5,2,1,4,2
+5,5,3,5,3,3,3,5,3,1,4,2
+5,5,2,5,2,2,2,6,4,2,4,2
+4,4,2,5,4,5,1,6,1,1,5,2
+6,5,5,6,4,4,2,4,2,1,6,2
+5,6,5,3,4,3,1,5,2,1,6,2
+4,0,0,5,0,0,0,4,0,0,0,3
+0,0,0,0,0,0,0,5,0,1,0,3
+2,2,2,4,2,2,3,2,2,1,0,3
+3,3,2,4,1,3,1,5,1,3,0,3
+2,2,1,5,1,2,1,6,0,1,1,3
+1,1,1,6,1,1,1,2,1,1,1,3
+3,1,1,5,1,2,1,2,1,1,1,3
+1,1,1,3,1,2,1,3,1,1,1,3
+1,3,1,4,1,3,2,3,1,1,1,3
+1,3,1,4,1,1,1,4,1,1,1,3
+3,2,2,5,3,2,1,4,1,1,1,3
+5,2,1,6,1,2,1,5,1,1,1,3
+1,1,1,4,2,4,1,5,1,1,1,3
+1,2,1,5,2,4,1,6,1,1,1,3
+2,2,2,4,2,2,1,1,2,1,1,3
+1,1,2,4,2,3,1,3,2,1,1,3
+3,1,1,6,1,4,1,3,2,1,1,3
+5,5,2,5,5,5,1,6,2,1,1,3
+3,1,1,1,1,1,3,1,3,1,1,3
+2,1,2,2,1,1,2,3,3,1,1,3
+2,1,1,3,2,2,2,2,1,2,1,3
+2,2,3,4,1,2,3,2,2,3,1,3
+2,2,1,3,2,2,1,2,1,1,2,3
+3,2,1,4,3,4,1,3,1,1,2,3
+1,2,1,3,2,2,2,3,1,1,2,3
+4,2,1,6,2,3,1,4,1,1,2,3
+2,4,1,3,3,3,1,5,1,1,2,3
+2,2,1,4,2,2,2,5,1,1,2,3
+2,2,1,5,2,2,2,5,1,1,2,3
+2,2,1,4,0,0,1,6,1,1,2,3
+1,1,1,5,2,2,1,6,1,1,2,3
+2,2,1,6,3,3,1,6,1,1,2,3
+3,3,1,4,2,4,1,6,1,1,2,3
+3,4,1,5,3,5,1,6,1,1,2,3
+2,2,2,2,2,2,1,2,2,1,2,3
+1,1,1,5,1,2,1,3,2,1,2,3
+2,3,1,5,2,4,3,3,3,1,2,3
+2,3,2,5,2,3,1,4,3,1,2,3
+2,2,0,4,2,2,2,5,3,1,2,3
+2,1,1,5,2,2,4,5,4,1,2,3
+1,1,1,5,1,2,3,2,1,2,2,3
+2,2,2,5,2,2,2,3,2,2,2,3
+2,0,2,4,2,3,1,2,1,1,3,3
+5,3,1,2,1,3,2,3,1,1,3,3
+1,1,1,5,3,1,1,4,1,1,3,3
+2,3,1,5,1,4,1,4,1,1,3,3
+2,4,1,5,2,3,1,5,1,1,3,3
+4,3,2,5,3,4,1,5,1,1,3,3
+4,3,1,3,2,5,1,5,1,1,3,3
+2,1,2,6,1,3,1,2,2,1,3,3
+2,2,2,6,0,4,0,4,2,1,3,3
+3,3,2,4,3,4,1,4,2,1,3,3
+3,2,1,5,4,0,1,6,2,1,3,3
+2,2,1,4,0,1,2,3,2,3,3,3
+3,3,3,3,3,3,3,3,3,3,3,3
+3,3,3,3,3,3,5,3,3,3,3,3
+3,3,3,3,3,3,3,3,4,3,3,3
+4,3,2,6,3,3,4,5,4,3,3,3
+4,3,1,4,3,3,1,5,1,1,4,3
+5,3,1,4,3,4,1,4,2,1,4,3
+4,5,2,6,5,5,3,4,2,1,4,3
+4,3,1,5,1,5,2,6,3,1,4,3
+5,3,3,4,4,4,1,6,4,1,4,3
+4,4,1,5,1,4,3,5,3,2,4,3
+5,3,3,4,3,3,2,4,4,2,4,3
+5,6,5,5,6,6,4,6,2,3,4,3
+3,4,2,6,2,6,5,4,2,5,4,3
+4,4,1,5,4,4,1,5,2,1,5,3
+6,4,6,3,5,4,2,6,4,2,6,3
+4,0,0,4,0,4,4,5,0,0,0,4
+2,1,1,6,1,3,1,3,1,1,0,4
+0,0,0,5,4,5,0,2,4,1,0,4
+1,1,1,5,1,1,1,5,0,1,1,4
+1,1,1,6,1,2,1,1,1,1,1,4
+2,2,1,5,1,2,2,1,1,1,1,4
+2,1,1,4,2,2,1,2,1,1,1,4
+1,1,1,6,1,3,1,2,1,1,1,4
+1,1,2,5,1,4,1,3,1,1,1,4
+1,1,1,6,2,3,1,4,1,1,1,4
+1,3,1,4,1,4,2,4,1,1,1,4
+3,2,2,5,3,4,2,4,1,1,1,4
+1,2,1,4,1,1,3,4,1,1,1,4
+1,1,1,4,1,2,1,5,1,1,1,4
+1,1,3,5,1,2,1,5,1,1,1,4
+2,3,1,5,3,5,2,6,1,1,1,4
+3,2,2,4,2,2,2,3,2,1,1,4
+2,3,2,3,2,4,2,4,2,1,1,4
+2,3,2,4,2,2,2,5,2,1,1,4
+2,2,2,1,2,2,2,3,3,1,1,4
+1,1,1,6,1,3,2,6,3,1,1,4
+4,1,2,6,4,1,5,2,4,1,1,4
+1,3,1,6,2,3,1,3,4,1,1,4
+1,1,2,1,1,1,2,1,4,2,1,4
+1,1,1,6,1,2,2,2,1,3,1,4
+1,1,1,5,2,1,3,4,4,3,1,4
+4,3,3,3,2,2,2,3,2,4,1,4
+2,3,2,6,1,1,4,4,4,5,1,4
+1,1,1,4,1,3,5,1,4,6,1,4
+5,2,1,5,1,4,5,1,1,1,2,4
+2,1,2,3,2,1,2,2,1,1,2,4
+2,2,1,6,1,2,1,3,1,1,2,4
+3,2,1,5,2,3,1,3,1,1,2,4
+4,3,1,3,3,4,1,4,1,1,2,4
+4,2,1,5,2,3,1,5,1,1,2,4
+3,2,1,5,1,4,1,5,1,1,2,4
+5,2,2,5,2,5,1,5,1,1,2,4
+1,3,1,4,1,2,1,6,1,1,2,4
+1,2,1,6,1,3,1,6,1,1,2,4
+2,2,2,3,3,3,1,6,1,1,2,4
+2,2,1,6,3,5,1,6,1,1,2,4
+3,3,2,3,4,3,4,6,1,1,2,4
+1,2,1,6,1,2,1,3,2,1,2,4
+2,2,2,6,3,3,1,4,2,1,2,4
+3,2,2,5,3,4,1,4,2,1,2,4
+2,2,1,4,3,2,2,4,2,1,2,4
+2,2,1,4,1,3,2,4,2,1,2,4
+2,2,1,6,2,2,3,4,2,1,2,4
+2,2,2,5,3,4,1,5,2,1,2,4
+2,2,1,5,2,2,2,4,3,1,2,4
+2,3,2,5,2,2,3,4,3,1,2,4
+3,3,1,4,2,6,3,5,3,1,2,4
+3,2,2,5,3,6,1,6,3,1,2,4
+3,2,1,5,2,3,2,6,3,1,2,4
+1,1,1,6,2,3,2,3,4,1,2,4
+1,3,3,5,1,3,3,4,4,1,2,4
+2,2,2,3,2,3,5,4,5,1,2,4
+3,3,1,4,2,4,5,4,1,2,2,4
+4,3,2,5,2,2,2,2,2,2,2,4
+4,2,2,5,3,3,3,2,2,2,2,4
+0,2,2,5,2,2,3,4,2,4,2,4
+1,1,1,2,1,1,3,2,1,1,3,4
+2,3,1,4,1,3,1,3,1,1,3,4
+4,3,1,6,4,3,2,5,1,1,3,4
+2,3,1,4,1,6,1,6,1,1,3,4
+3,1,1,2,3,1,1,5,2,1,3,4
+4,0,0,5,3,0,3,3,3,1,3,4
+3,2,1,6,3,3,1,6,1,2,3,4
+1,1,1,5,3,3,2,3,2,2,3,4
+4,4,4,5,1,1,1,2,5,5,3,4
+4,3,1,6,2,5,2,3,1,1,4,4
+5,3,1,5,1,2,1,5,1,1,4,4
+1,4,1,6,1,4,4,6,1,1,4,4
+4,2,1,5,3,2,2,4,4,1,4,4
+5,4,1,4,4,3,1,5,5,1,4,4
+3,3,2,4,3,0,3,4,0,4,4,4
+3,2,2,5,2,2,2,4,2,4,4,4
+4,3,2,5,1,3,2,6,4,4,4,4
+5,4,4,5,5,4,2,6,3,1,5,4
+4,4,2,5,5,5,2,6,4,1,5,4
+0,0,0,0,0,0,4,3,0,0,0,5
+0,0,0,0,0,4,0,4,0,0,0,5
+1,1,1,2,1,3,3,2,4,1,0,5
+0,0,0,2,0,0,0,3,0,4,0,5
+1,1,1,3,1,1,1,1,1,1,1,5
+1,1,1,4,1,1,2,1,1,1,1,5
+1,1,1,4,1,1,1,2,1,1,1,5
+1,1,1,4,1,1,1,2,1,1,1,5
+1,1,1,6,1,1,1,2,1,1,1,5
+1,1,1,2,1,3,3,2,1,1,1,5
+1,1,1,6,1,1,1,3,1,1,1,5
+1,1,1,3,1,2,1,3,1,1,1,5
+1,1,1,3,1,2,1,3,1,1,1,5
+3,1,1,5,2,2,1,3,1,1,1,5
+1,1,1,3,1,2,4,3,1,1,1,5
+1,1,1,6,1,4,4,3,1,1,1,5
+1,1,1,6,1,2,1,4,1,1,1,5
+2,2,1,4,2,2,2,4,1,1,1,5
+1,2,1,5,1,3,3,4,1,1,1,5
+2,2,2,4,1,2,1,5,1,1,1,5
+1,2,1,5,2,4,2,5,1,1,1,5
+4,3,1,5,4,4,1,6,1,1,1,5
+1,2,2,6,2,2,1,2,2,1,1,5
+0,1,1,3,2,1,3,2,2,1,1,5
+1,2,2,3,1,2,1,3,2,1,1,5
+1,1,1,4,1,1,2,3,2,1,1,5
+4,2,1,6,1,3,2,3,2,1,1,5
+1,1,2,3,1,2,1,4,2,1,1,5
+1,1,1,5,1,3,1,4,2,1,1,5
+1,2,1,5,2,3,1,4,2,1,1,5
+2,2,2,5,2,2,2,4,2,1,1,5
+1,2,1,6,2,2,1,5,2,1,1,5
+1,2,2,4,2,3,3,1,3,1,1,5
+2,2,1,5,2,3,2,4,3,1,1,5
+1,1,1,2,1,1,1,3,4,1,1,5
+1,2,2,5,1,2,2,4,4,1,1,5
+1,2,1,5,1,2,1,2,1,2,1,5
+1,1,1,3,3,3,2,2,1,2,1,5
+1,2,1,4,2,2,3,4,2,2,1,5
+1,2,1,4,3,3,3,4,2,2,1,5
+1,1,1,4,1,1,5,2,0,3,1,5
+2,1,1,2,1,3,4,1,1,3,1,5
+1,1,1,4,1,1,6,1,1,3,1,5
+1,1,1,4,1,2,2,2,1,3,1,5
+2,2,1,5,4,2,2,2,1,3,1,5
+2,1,1,3,1,1,3,2,1,3,1,5
+3,3,1,2,1,1,1,4,1,3,1,5
+1,1,1,3,1,2,4,5,4,3,1,5
+2,1,2,1,1,1,4,1,1,4,1,5
+1,1,1,2,1,1,1,3,1,4,1,5
+2,1,1,3,1,2,2,1,2,4,1,5
+1,1,2,5,1,2,3,3,3,4,1,5
+1,1,1,2,1,1,3,1,1,5,1,5
+1,1,1,5,1,1,5,1,1,5,1,5
+2,2,2,5,4,1,5,1,2,5,1,5
+1,1,1,0,1,1,1,1,1,6,1,5
+1,1,1,1,1,1,1,1,1,6,1,5
+1,1,1,5,1,1,5,1,1,6,1,5
+5,4,0,6,3,2,0,5,0,1,2,5
+4,4,1,6,1,3,4,2,1,1,2,5
+2,2,1,5,2,2,1,3,1,1,2,5
+3,2,2,3,1,1,4,3,1,1,2,5
+5,3,2,4,3,2,1,5,1,1,2,5
+1,3,1,3,4,2,1,5,1,1,2,5
+1,3,1,3,1,3,1,5,1,1,2,5
+1,2,1,6,1,3,1,5,1,1,2,5
+2,3,1,6,1,4,2,5,1,1,2,5
+2,4,1,5,2,4,2,5,1,1,2,5
+4,3,1,5,2,3,4,6,1,1,2,5
+1,1,1,4,1,3,2,3,2,1,2,5
+2,2,1,6,2,3,2,4,2,1,2,5
+4,1,2,6,4,4,4,1,3,1,2,5
+2,2,2,3,2,2,1,2,4,1,2,5
+1,2,1,5,2,1,2,4,1,2,2,5
+2,2,1,5,2,2,3,3,1,3,2,5
+0,3,1,5,1,1,3,4,1,3,2,5
+2,2,2,4,2,2,3,2,2,3,2,5
+3,1,1,4,1,1,1,1,3,3,2,5
+3,2,1,3,1,4,5,3,1,4,2,5
+3,2,1,3,1,2,1,4,1,4,2,5
+2,2,2,6,2,3,3,4,2,4,2,5
+4,2,1,3,2,3,5,2,3,4,2,5
+1,2,1,3,2,2,3,4,3,4,2,5
+3,3,3,3,3,3,4,3,4,4,2,5
+2,2,2,3,2,2,5,3,2,5,2,5
+2,2,2,5,2,5,5,3,2,5,2,5
+1,1,1,5,1,1,1,1,1,1,3,5
+4,3,2,5,2,3,1,5,1,1,3,5
+3,4,1,6,4,5,1,5,1,1,3,5
+4,4,1,6,5,5,1,5,1,1,3,5
+2,2,2,6,1,2,3,2,2,1,3,5
+1,2,1,5,1,3,4,3,2,1,3,5
+4,4,2,5,1,3,4,4,2,1,3,5
+4,2,1,5,1,2,1,5,2,1,3,5
+4,3,1,6,2,3,1,5,2,1,3,5
+3,3,1,6,3,4,1,5,2,1,3,5
+2,3,1,4,1,3,3,5,2,1,3,5
+4,3,3,6,2,3,4,3,3,1,3,5
+2,1,1,4,3,3,3,4,3,1,3,5
+3,1,3,4,3,3,4,4,3,3,3,5
+2,3,1,4,1,1,3,5,3,3,3,5
+1,1,1,2,3,1,4,1,1,4,3,5
+3,2,1,6,2,2,2,4,3,4,3,5
+4,3,2,5,3,4,4,3,5,4,3,5
+5,5,2,5,3,5,4,5,3,5,3,5
+2,5,1,6,1,3,1,2,4,5,3,5
+4,4,1,4,1,2,1,1,1,1,4,5
+2,3,1,6,2,3,1,4,1,1,4,5
+2,2,1,6,1,2,3,5,1,1,4,5
+3,4,1,5,2,3,3,5,1,1,4,5
+4,5,2,5,3,5,1,6,2,1,4,5
+5,4,1,5,3,5,1,4,3,1,4,5
+4,4,3,6,3,3,4,4,3,1,4,5
+3,5,3,4,4,4,4,4,3,5,4,5
+3,1,1,4,3,4,1,4,1,1,5,5
+5,5,1,6,2,5,3,4,1,1,5,5
+4,2,1,6,4,1,1,3,3,1,5,5
+3,1,1,6,1,4,5,1,1,5,5,5
+5,4,3,4,5,5,4,5,2,5,5,5
+6,2,2,4,4,4,5,4,6,5,6,5
+0,0,0,0,0,0,0,0,0,0,0,6
+0,0,0,0,0,0,0,0,0,0,0,6
+0,0,0,4,0,0,0,0,0,0,0,6
+4,1,3,0,0,5,0,2,0,0,0,6
+0,0,1,2,2,2,1,1,1,1,0,6
+4,3,2,3,1,2,2,2,2,1,0,6
+2,0,0,0,0,0,6,5,0,5,0,6
+4,3,2,5,3,3,5,3,3,6,0,6
+1,1,1,1,1,1,1,1,1,1,1,6
+1,1,1,2,1,1,1,1,1,1,1,6
+2,1,1,5,3,3,1,1,1,1,1,6
+0,1,1,5,2,2,2,1,1,1,1,6
+1,2,2,2,3,3,2,1,1,1,1,6
+1,1,1,1,1,1,4,1,1,1,1,6
+1,1,1,1,1,1,1,2,1,1,1,6
+1,1,1,4,1,1,1,2,1,1,1,6
+1,1,1,6,1,1,1,2,1,1,1,6
+1,1,1,4,1,2,2,2,1,1,1,6
+1,1,1,4,1,2,3,2,1,1,1,6
+1,1,2,3,2,2,3,2,1,1,1,6
+2,1,1,5,2,2,1,3,1,1,1,6
+1,1,1,2,1,2,3,3,1,1,1,6
+1,2,1,1,4,3,3,3,1,1,1,6
+1,1,1,5,1,2,1,4,1,1,1,6
+1,2,2,5,2,2,3,4,1,1,1,6
+1,1,1,5,3,3,5,4,1,1,1,6
+1,1,1,4,1,1,1,5,1,1,1,6
+1,1,1,5,1,1,6,5,1,1,1,6
+3,1,1,4,1,1,1,6,1,1,1,6
+1,1,1,4,1,2,1,4,2,1,1,6
+1,2,1,3,1,1,2,4,2,1,1,6
+1,2,1,5,1,2,4,4,2,1,1,6
+2,2,1,6,2,4,1,6,2,1,1,6
+1,1,1,4,1,1,2,2,3,1,1,6
+3,2,1,5,2,2,1,3,3,1,1,6
+1,2,1,5,1,3,5,4,3,1,1,6
+6,3,1,6,6,1,6,6,3,1,1,6
+3,2,3,5,1,3,3,3,4,1,1,6
+4,1,1,5,1,1,2,2,5,1,1,6
+1,1,1,2,1,1,2,1,1,2,1,6
+2,1,1,6,5,1,6,2,1,2,1,6
+2,2,2,6,5,2,4,1,2,2,1,6
+1,1,1,5,1,1,1,2,1,3,1,6
+1,1,1,4,1,1,3,4,1,3,1,6
+1,1,2,6,3,3,3,4,2,3,1,6
+1,2,1,5,3,3,5,3,4,3,1,6
+2,1,1,5,1,1,1,1,1,4,1,6
+1,1,1,2,1,1,4,1,1,4,1,6
+1,1,1,2,1,1,5,2,2,4,1,6
+1,2,1,4,1,3,0,4,2,4,1,6
+2,1,1,4,1,1,5,3,3,4,1,6
+1,1,1,4,2,2,6,4,3,4,1,6
+1,1,1,1,1,1,1,1,1,5,1,6
+1,1,1,3,1,1,3,1,1,5,1,6
+1,1,1,1,1,1,5,1,1,5,1,6
+1,1,1,3,2,2,5,1,2,5,1,6
+1,1,1,3,2,1,5,2,5,5,1,6
+1,1,1,6,4,4,6,1,1,6,1,6
+2,1,1,2,1,2,1,1,1,1,2,6
+2,1,1,6,5,5,2,1,1,1,2,6
+1,1,2,2,1,1,5,1,1,1,2,6
+2,3,1,4,1,1,1,3,1,1,2,6
+1,0,1,4,2,5,2,4,1,1,2,6
+1,3,1,5,1,3,2,5,1,1,2,6
+3,3,1,5,2,2,3,5,1,1,2,6
+0,2,2,3,1,4,3,4,2,1,2,6
+2,2,2,5,1,2,2,5,2,1,2,6
+3,3,3,4,1,3,1,4,3,1,2,6
+4,2,3,5,3,4,2,4,3,1,2,6
+0,1,1,5,1,2,2,1,4,1,2,6
+3,1,1,2,1,2,2,2,1,2,2,6
+1,2,1,5,1,2,3,4,2,2,2,6
+1,3,1,5,2,3,3,3,1,3,2,6
+1,1,2,3,1,2,3,2,2,4,2,6
+2,1,1,5,1,4,1,5,2,4,2,6
+1,2,3,4,1,3,5,2,1,5,2,6
+2,2,2,4,2,1,6,3,2,5,2,6
+5,1,1,2,1,1,6,2,5,5,2,6
+1,2,0,5,2,3,6,1,0,1,3,6
+3,2,2,5,2,2,2,4,2,1,3,6
+2,4,2,6,2,3,4,5,2,1,3,6
+2,3,3,5,1,3,5,2,2,2,3,6
+2,1,1,5,1,1,4,3,1,3,3,6
+2,2,1,5,2,2,2,5,1,3,3,6
+2,3,1,4,4,4,4,5,1,3,3,6
+3,2,1,5,3,3,2,6,1,3,3,6
+2,2,1,5,1,3,4,3,2,3,3,6
+4,2,1,5,2,3,3,1,1,4,3,6
+3,4,3,6,2,6,3,5,3,4,3,6
+3,3,3,4,3,3,4,5,3,4,3,6
+4,4,4,5,4,4,5,6,4,4,3,6
+4,3,2,5,2,3,4,3,1,6,3,6
+2,3,1,4,1,3,3,1,1,1,4,6
+4,5,2,5,1,4,4,5,1,1,4,6
+4,1,1,6,1,2,2,3,1,2,4,6
+4,2,1,5,4,3,5,3,2,4,4,6
+2,3,3,4,2,3,4,1,2,5,4,6
+1,1,2,5,3,4,1,2,2,1,5,6
+5,4,3,4,5,4,1,4,3,4,5,6
+5,3,4,2,3,4,5,5,4,5,5,6
+6,6,6,6,6,6,6,6,1,6,6,6
+    """.trimIndent()
+    val df = DataFrame.readCsvStr(csvString)
+
+    val poll = Poll(
+        id = 42,
+        pollConfig = PollConfig(
+            subject = "2007",
+            proposals = df.columnNames(),
+            grading = Grading.Quality7Grading,
+        ),
+        ballots = df.map {
+            Ballot(
+                judgments = it.values().mapIndexed { index, value ->
+                    Judgment(
+                        proposal = index,
+                        grade = value.toString().toInt(),
+                    )
+                }.toPersistentList()
+            )
+        },
+    )
+    
+    val analyzer = ProximityAnalyzer()
+    val analysis = analyzer.analyze(
+        poll = poll,
+    )
+
+    JmTheme {
+        Column(modifier) {
+            Text("")
             ProximitySpider(
                 modifier = Modifier.height(400.dp),
                 analysis = analysis,
