+++ app/src/main/java/com/illiouchine/jm/ui/composable/plot/lib/SpiderChart.kt
@@ -187,655 +187,3 @@ fun PreviewDynamicSpiderChart(modifier: Modifier = Modifier) {
         )
     }
 }
-
-
-@Preview(
-    name = "2007 Phone (Portrait)",
-    showSystemUi = false,
-    uiMode = Configuration.UI_MODE_NIGHT_YES,
-    fontScale = 1.0f,
-)
-@Composable
-fun PreviewDynamicHugeSpiderChart(modifier: Modifier = Modifier) {
-    val csvString = """
-Besancenot,Buffet,Schivardi,Bayrou,Bové,Voynet,De Villiers,Royal,Nihous,Le Pen,Laguiller,Sarkozy
-2,0,0,0,0,0,0,0,0,0,0,0
-0,0,0,5,0,0,0,0,0,0,0,0
-0,0,0,0,0,0,0,6,0,0,0,0
-0,0,0,0,0,0,0,6,0,0,0,0
-5,5,0,0,0,0,0,6,0,0,0,0
-2,3,2,5,1,3,2,5,4,2,0,0
-1,1,0,0,0,0,1,0,0,1,1,0
-3,1,1,4,1,1,1,2,1,1,1,0
-3,3,2,6,5,4,1,4,3,1,3,0
-3,4,3,6,2,4,0,4,0,0,6,0
-3,0,0,0,2,1,0,4,0,1,0,1
-4,6,2,2,2,2,1,5,1,1,0,1
-5,5,2,2,1,2,1,2,1,4,0,1
-3,2,0,2,2,6,0,6,0,1,1,1
-1,1,1,1,1,1,1,1,1,1,1,1
-1,1,1,1,1,1,1,1,1,1,1,1
-1,1,1,1,1,1,1,1,1,1,1,1
-4,6,3,2,3,1,1,1,1,1,1,1
-4,4,3,2,4,1,1,1,1,1,1,1
-4,4,1,4,3,4,1,1,1,1,1,1
-1,2,3,4,4,2,1,2,1,1,1,1
-4,5,1,2,4,2,1,3,1,1,1,1
-2,1,1,2,2,3,1,3,1,1,1,1
-2,1,1,6,3,4,1,3,1,1,1,1
-1,1,1,6,1,1,1,4,1,1,1,1
-5,4,1,5,2,1,1,4,1,1,1,1
-5,2,1,5,1,2,1,4,1,1,1,1
-3,1,1,5,2,2,1,4,1,1,1,1
-1,1,1,5,1,3,1,4,1,1,1,1
-1,1,2,3,2,3,1,4,1,1,1,1
-1,1,1,6,3,4,1,4,1,1,1,1
-6,5,3,2,1,1,1,5,1,1,1,1
-1,1,1,3,1,2,1,5,1,1,1,1
-1,2,1,5,2,2,1,5,1,1,1,1
-2,6,1,3,2,4,1,5,1,1,1,1
-1,1,1,6,1,5,1,5,1,1,1,1
-1,1,1,6,1,5,1,5,1,1,1,1
-1,1,1,1,1,1,1,6,1,1,1,1
-1,1,1,3,1,1,1,6,1,1,1,1
-4,1,1,5,1,1,1,6,1,1,1,1
-1,1,1,5,2,1,1,6,1,1,1,1
-3,5,2,4,2,3,1,6,1,1,1,1
-5,3,1,3,3,3,1,6,1,1,1,1
-3,2,1,1,1,4,1,6,1,1,1,1
-4,6,3,3,3,6,1,6,1,1,1,1
-2,2,1,4,1,2,2,6,1,1,1,1
-2,2,1,5,2,2,1,1,2,1,1,1
-3,3,3,5,2,2,2,3,2,1,1,1
-2,2,1,5,2,3,1,5,2,1,1,1
-3,2,3,3,1,2,3,5,2,1,1,1
-4,4,3,4,5,2,1,6,2,1,1,1
-3,3,2,6,2,5,1,6,2,1,1,1
-4,2,1,3,4,4,1,6,3,1,1,1
-2,2,2,4,1,3,0,6,4,1,1,1
-1,3,1,4,2,3,1,6,4,1,1,1
-5,3,2,6,1,1,1,2,6,1,1,1
-1,1,1,1,1,1,1,1,1,6,1,1
-4,5,0,3,5,3,1,0,0,1,2,1
-2,5,2,3,2,4,1,6,0,1,2,1
-3,1,5,3,2,2,1,1,1,1,2,1
-4,6,1,1,4,3,1,3,1,1,2,1
-4,3,2,2,2,4,1,3,1,1,2,1
-2,2,1,2,2,3,1,4,1,1,2,1
-2,2,1,4,2,3,1,4,1,1,2,1
-3,2,1,4,2,3,1,4,1,1,2,1
-3,2,1,4,2,4,1,4,1,1,2,1
-2,2,2,3,3,4,1,4,1,1,2,1
-5,5,2,2,3,6,1,4,1,1,2,1
-1,2,1,4,1,1,1,5,1,1,2,1
-2,3,1,4,1,2,1,5,1,1,2,1
-2,3,2,4,2,2,1,5,1,1,2,1
-2,4,1,4,2,4,1,5,1,1,2,1
-3,4,1,5,2,4,1,5,1,1,2,1
-6,4,3,1,6,5,1,5,1,1,2,1
-2,2,4,4,2,6,1,5,1,1,2,1
-2,2,1,5,2,1,1,6,1,1,2,1
-2,3,1,2,2,2,1,6,1,1,2,1
-2,2,2,3,2,2,1,6,1,1,2,1
-2,3,1,4,2,3,1,6,1,1,2,1
-5,4,2,1,5,3,1,6,1,1,2,1
-2,2,1,5,3,4,1,6,1,1,2,1
-2,3,1,5,2,5,2,6,1,1,2,1
-2,2,2,2,2,2,1,2,2,1,2,1
-2,2,2,5,3,2,1,4,2,1,2,1
-2,3,2,6,2,3,1,4,2,1,2,1
-3,2,2,6,3,3,1,4,2,1,2,1
-2,2,1,4,1,2,1,5,2,1,2,1
-2,2,1,6,2,2,1,5,2,1,2,1
-3,4,2,5,4,3,1,5,2,1,2,1
-6,4,2,4,5,3,1,5,2,1,2,1
-2,3,2,4,2,4,1,5,2,1,2,1
-4,4,2,3,4,4,1,5,2,1,2,1
-2,3,0,5,1,5,1,5,2,1,2,1
-4,4,2,4,2,5,1,5,2,1,2,1
-3,2,2,4,5,6,0,6,2,1,2,1
-6,6,2,2,2,2,1,6,2,1,2,1
-2,2,1,2,1,3,1,6,2,1,2,1
-3,3,3,2,3,3,1,6,2,1,2,1
-2,2,2,3,3,3,1,6,2,1,2,1
-3,5,2,3,4,5,1,6,2,1,2,1
-3,3,2,5,3,4,2,6,2,1,2,1
-6,3,2,6,4,5,3,6,2,1,2,1
-2,2,2,4,1,3,2,2,3,1,2,1
-4,5,3,6,4,5,2,4,3,1,2,1
-5,4,1,5,1,1,1,5,3,1,2,1
-3,4,1,2,3,2,1,5,4,1,2,1
-1,5,3,4,2,3,4,3,2,5,2,1
-2,3,1,1,6,2,1,1,1,1,3,1
-4,4,2,5,4,1,1,2,1,1,3,1
-6,5,3,2,0,4,1,2,1,1,3,1
-3,3,2,4,4,4,1,2,1,1,3,1
-5,2,2,1,4,6,1,2,1,1,3,1
-5,3,1,5,5,6,1,2,1,1,3,1
-5,5,1,1,2,3,1,3,1,1,3,1
-3,4,1,6,3,3,1,4,1,1,3,1
-3,3,2,4,3,4,1,4,1,1,3,1
-4,4,2,2,4,5,1,4,1,1,3,1
-3,4,3,4,3,3,1,5,1,1,3,1
-4,4,1,5,3,5,1,5,1,1,3,1
-3,6,1,3,1,6,1,5,1,1,3,1
-4,5,2,4,2,3,2,5,1,1,3,1
-4,3,2,5,4,6,2,5,1,1,3,1
-3,3,3,3,2,1,1,6,1,1,3,1
-4,3,1,3,3,3,1,6,1,1,3,1
-4,5,2,5,4,3,1,6,1,1,3,1
-4,4,0,4,4,4,1,6,1,1,3,1
-5,4,1,4,4,5,1,6,1,1,3,1
-3,3,1,5,3,3,2,6,1,1,3,1
-3,4,2,6,3,3,2,6,1,1,3,1
-5,4,3,5,3,3,1,3,2,1,3,1
-4,3,2,1,3,2,1,4,2,1,3,1
-4,2,1,4,3,3,1,5,2,1,3,1
-2,4,1,3,2,4,1,5,2,1,3,1
-5,3,1,6,3,5,1,5,2,1,3,1
-4,3,2,5,1,5,1,6,2,1,3,1
-4,4,1,5,2,5,1,6,2,1,3,1
-4,3,2,5,3,3,2,6,2,1,3,1
-3,3,1,6,1,5,1,2,3,1,3,1
-4,6,3,5,5,5,1,6,3,1,3,1
-2,4,4,4,5,5,1,6,4,1,3,1
-5,0,0,5,0,0,0,6,0,1,4,1
-2,2,1,4,2,2,1,5,1,1,4,1
-4,3,1,5,3,3,1,5,1,1,4,1
-6,5,4,3,4,3,1,5,1,1,4,1
-6,5,2,3,4,5,1,5,1,1,4,1
-6,6,2,1,5,5,1,5,1,1,4,1
-5,6,4,4,4,6,1,5,1,1,4,1
-3,5,5,3,2,2,1,6,1,1,4,1
-2,5,1,5,1,4,1,6,1,1,4,1
-4,5,2,4,2,4,1,6,1,1,4,1
-4,4,2,3,3,4,1,6,1,1,4,1
-5,4,2,3,2,6,1,3,2,1,4,1
-1,2,1,5,3,2,5,3,2,1,4,1
-6,6,4,3,4,5,1,5,2,1,4,1
-6,4,2,3,4,3,1,6,2,1,4,1
-5,5,1,3,4,4,1,6,2,1,4,1
-5,5,2,2,5,5,1,6,2,1,4,1
-4,4,1,3,3,6,1,6,2,1,4,1
-5,5,2,4,4,4,2,6,2,1,4,1
-4,5,3,6,6,6,2,6,2,1,4,1
-5,4,3,5,5,5,1,4,3,1,4,1
-5,5,3,5,2,3,1,6,3,1,4,1
-6,5,3,2,1,4,1,6,3,1,4,1
-1,4,3,6,4,5,4,6,3,1,4,1
-2,4,2,5,3,3,1,5,5,1,4,1
-4,6,4,1,4,4,1,4,0,1,5,1
-6,0,0,5,0,0,0,6,0,1,5,1
-6,5,2,2,4,3,1,2,1,1,5,1
-4,4,4,4,4,4,1,5,1,1,5,1
-5,6,1,5,5,5,1,5,1,1,5,1
-3,3,1,5,2,2,1,6,1,1,5,1
-5,5,1,2,3,4,1,6,1,1,5,1
-5,5,0,4,6,4,1,3,2,1,5,1
-5,5,3,3,6,6,1,5,2,1,5,1
-5,3,1,5,4,2,1,6,2,1,5,1
-6,4,3,6,5,4,1,6,2,1,5,1
-5,4,3,6,0,5,2,5,3,1,5,1
-4,4,3,6,5,5,5,5,4,1,5,1
-5,6,4,6,5,4,1,6,1,2,5,1
-2,3,1,4,2,3,3,2,1,5,5,1
-0,6,0,5,0,0,0,6,0,1,6,1
-6,4,1,1,1,1,1,1,1,1,6,1
-5,3,2,5,5,3,1,5,2,1,6,1
-6,5,3,4,5,5,1,5,2,1,6,1
-6,4,3,1,5,5,1,4,4,1,6,1
-5,6,4,6,5,5,1,5,5,5,6,1
-0,0,0,0,0,0,0,2,0,1,0,2
-1,4,2,3,4,1,0,6,0,1,1,2
-1,1,1,5,1,3,1,2,1,1,1,2
-4,5,1,4,4,4,1,2,1,1,1,2
-2,3,1,4,2,1,1,3,1,1,1,2
-2,2,1,5,2,2,1,4,1,1,1,2
-1,4,1,5,2,3,1,5,1,1,1,2
-4,4,1,3,2,4,1,5,1,1,1,2
-1,2,1,4,4,4,1,5,1,1,1,2
-1,2,1,5,1,5,1,5,1,1,1,2
-1,1,1,2,3,3,6,6,1,1,1,2
-2,2,2,2,2,2,1,3,2,1,1,2
-1,1,2,5,1,1,1,3,3,1,1,2
-2,2,2,5,3,3,1,2,4,1,1,2
-2,1,3,2,1,1,1,2,2,2,1,2
-2,2,1,4,2,2,1,3,1,1,2,2
-3,2,1,5,2,2,1,5,1,1,2,2
-3,3,2,3,2,3,1,5,1,1,2,2
-3,2,1,6,2,3,1,5,1,1,2,2
-2,2,1,4,2,4,1,5,1,1,2,2
-1,1,1,5,1,3,1,6,1,1,2,2
-3,3,2,4,3,5,1,6,1,1,2,2
-2,2,2,2,2,2,1,2,2,1,2,2
-2,2,2,5,2,2,1,4,2,1,2,2
-2,2,1,2,2,3,1,4,2,1,2,2
-2,3,1,5,1,2,1,5,2,1,2,2
-3,2,2,4,2,3,1,5,2,1,2,2
-3,3,3,3,2,2,1,5,3,2,2,2
-4,4,2,6,4,2,2,6,4,3,2,2
-2,2,1,3,2,2,1,2,1,1,3,2
-4,3,1,5,4,4,1,3,1,1,3,2
-1,3,1,4,1,3,1,5,1,1,3,2
-5,6,2,3,1,6,1,5,1,1,3,2
-3,3,2,6,1,4,1,3,2,1,3,2
-5,4,2,4,5,6,1,6,3,1,3,2
-3,3,2,6,1,1,1,3,4,1,3,2
-4,5,1,3,3,4,1,5,1,1,4,2
-3,4,2,5,4,4,3,4,2,1,4,2
-5,4,3,3,3,4,1,5,2,1,4,2
-4,5,2,4,4,4,1,5,2,1,4,2
-5,5,3,5,3,3,3,5,3,1,4,2
-5,5,2,5,2,2,2,6,4,2,4,2
-4,4,2,5,4,5,1,6,1,1,5,2
-6,5,5,6,4,4,2,4,2,1,6,2
-5,6,5,3,4,3,1,5,2,1,6,2
-4,0,0,5,0,0,0,4,0,0,0,3
-0,0,0,0,0,0,0,5,0,1,0,3
-2,2,2,4,2,2,3,2,2,1,0,3
-3,3,2,4,1,3,1,5,1,3,0,3
-2,2,1,5,1,2,1,6,0,1,1,3
-1,1,1,6,1,1,1,2,1,1,1,3
-3,1,1,5,1,2,1,2,1,1,1,3
-1,1,1,3,1,2,1,3,1,1,1,3
-1,3,1,4,1,3,2,3,1,1,1,3
-1,3,1,4,1,1,1,4,1,1,1,3
-3,2,2,5,3,2,1,4,1,1,1,3
-5,2,1,6,1,2,1,5,1,1,1,3
-1,1,1,4,2,4,1,5,1,1,1,3
-1,2,1,5,2,4,1,6,1,1,1,3
-2,2,2,4,2,2,1,1,2,1,1,3
-1,1,2,4,2,3,1,3,2,1,1,3
-3,1,1,6,1,4,1,3,2,1,1,3
-5,5,2,5,5,5,1,6,2,1,1,3
-3,1,1,1,1,1,3,1,3,1,1,3
-2,1,2,2,1,1,2,3,3,1,1,3
-2,1,1,3,2,2,2,2,1,2,1,3
-2,2,3,4,1,2,3,2,2,3,1,3
-2,2,1,3,2,2,1,2,1,1,2,3
-3,2,1,4,3,4,1,3,1,1,2,3
-1,2,1,3,2,2,2,3,1,1,2,3
-4,2,1,6,2,3,1,4,1,1,2,3
-2,4,1,3,3,3,1,5,1,1,2,3
-2,2,1,4,2,2,2,5,1,1,2,3
-2,2,1,5,2,2,2,5,1,1,2,3
-2,2,1,4,0,0,1,6,1,1,2,3
-1,1,1,5,2,2,1,6,1,1,2,3
-2,2,1,6,3,3,1,6,1,1,2,3
-3,3,1,4,2,4,1,6,1,1,2,3
-3,4,1,5,3,5,1,6,1,1,2,3
-2,2,2,2,2,2,1,2,2,1,2,3
-1,1,1,5,1,2,1,3,2,1,2,3
-2,3,1,5,2,4,3,3,3,1,2,3
-2,3,2,5,2,3,1,4,3,1,2,3
-2,2,0,4,2,2,2,5,3,1,2,3
-2,1,1,5,2,2,4,5,4,1,2,3
-1,1,1,5,1,2,3,2,1,2,2,3
-2,2,2,5,2,2,2,3,2,2,2,3
-2,0,2,4,2,3,1,2,1,1,3,3
-5,3,1,2,1,3,2,3,1,1,3,3
-1,1,1,5,3,1,1,4,1,1,3,3
-2,3,1,5,1,4,1,4,1,1,3,3
-2,4,1,5,2,3,1,5,1,1,3,3
-4,3,2,5,3,4,1,5,1,1,3,3
-4,3,1,3,2,5,1,5,1,1,3,3
-2,1,2,6,1,3,1,2,2,1,3,3
-2,2,2,6,0,4,0,4,2,1,3,3
-3,3,2,4,3,4,1,4,2,1,3,3
-3,2,1,5,4,0,1,6,2,1,3,3
-2,2,1,4,0,1,2,3,2,3,3,3
-3,3,3,3,3,3,3,3,3,3,3,3
-3,3,3,3,3,3,5,3,3,3,3,3
-3,3,3,3,3,3,3,3,4,3,3,3
-4,3,2,6,3,3,4,5,4,3,3,3
-4,3,1,4,3,3,1,5,1,1,4,3
-5,3,1,4,3,4,1,4,2,1,4,3
-4,5,2,6,5,5,3,4,2,1,4,3
-4,3,1,5,1,5,2,6,3,1,4,3
-5,3,3,4,4,4,1,6,4,1,4,3
-4,4,1,5,1,4,3,5,3,2,4,3
-5,3,3,4,3,3,2,4,4,2,4,3
-5,6,5,5,6,6,4,6,2,3,4,3
-3,4,2,6,2,6,5,4,2,5,4,3
-4,4,1,5,4,4,1,5,2,1,5,3
-6,4,6,3,5,4,2,6,4,2,6,3
-4,0,0,4,0,4,4,5,0,0,0,4
-2,1,1,6,1,3,1,3,1,1,0,4
-0,0,0,5,4,5,0,2,4,1,0,4
-1,1,1,5,1,1,1,5,0,1,1,4
-1,1,1,6,1,2,1,1,1,1,1,4
-2,2,1,5,1,2,2,1,1,1,1,4
-2,1,1,4,2,2,1,2,1,1,1,4
-1,1,1,6,1,3,1,2,1,1,1,4
-1,1,2,5,1,4,1,3,1,1,1,4
-1,1,1,6,2,3,1,4,1,1,1,4
-1,3,1,4,1,4,2,4,1,1,1,4
-3,2,2,5,3,4,2,4,1,1,1,4
-1,2,1,4,1,1,3,4,1,1,1,4
-1,1,1,4,1,2,1,5,1,1,1,4
-1,1,3,5,1,2,1,5,1,1,1,4
-2,3,1,5,3,5,2,6,1,1,1,4
-3,2,2,4,2,2,2,3,2,1,1,4
-2,3,2,3,2,4,2,4,2,1,1,4
-2,3,2,4,2,2,2,5,2,1,1,4
-2,2,2,1,2,2,2,3,3,1,1,4
-1,1,1,6,1,3,2,6,3,1,1,4
-4,1,2,6,4,1,5,2,4,1,1,4
-1,3,1,6,2,3,1,3,4,1,1,4
-1,1,2,1,1,1,2,1,4,2,1,4
-1,1,1,6,1,2,2,2,1,3,1,4
-1,1,1,5,2,1,3,4,4,3,1,4
-4,3,3,3,2,2,2,3,2,4,1,4
-2,3,2,6,1,1,4,4,4,5,1,4
-1,1,1,4,1,3,5,1,4,6,1,4
-5,2,1,5,1,4,5,1,1,1,2,4
-2,1,2,3,2,1,2,2,1,1,2,4
-2,2,1,6,1,2,1,3,1,1,2,4
-3,2,1,5,2,3,1,3,1,1,2,4
-4,3,1,3,3,4,1,4,1,1,2,4
-4,2,1,5,2,3,1,5,1,1,2,4
-3,2,1,5,1,4,1,5,1,1,2,4
-5,2,2,5,2,5,1,5,1,1,2,4
-1,3,1,4,1,2,1,6,1,1,2,4
-1,2,1,6,1,3,1,6,1,1,2,4
-2,2,2,3,3,3,1,6,1,1,2,4
-2,2,1,6,3,5,1,6,1,1,2,4
-3,3,2,3,4,3,4,6,1,1,2,4
-1,2,1,6,1,2,1,3,2,1,2,4
-2,2,2,6,3,3,1,4,2,1,2,4
-3,2,2,5,3,4,1,4,2,1,2,4
-2,2,1,4,3,2,2,4,2,1,2,4
-2,2,1,4,1,3,2,4,2,1,2,4
-2,2,1,6,2,2,3,4,2,1,2,4
-2,2,2,5,3,4,1,5,2,1,2,4
-2,2,1,5,2,2,2,4,3,1,2,4
-2,3,2,5,2,2,3,4,3,1,2,4
-3,3,1,4,2,6,3,5,3,1,2,4
-3,2,2,5,3,6,1,6,3,1,2,4
-3,2,1,5,2,3,2,6,3,1,2,4
-1,1,1,6,2,3,2,3,4,1,2,4
-1,3,3,5,1,3,3,4,4,1,2,4
-2,2,2,3,2,3,5,4,5,1,2,4
-3,3,1,4,2,4,5,4,1,2,2,4
-4,3,2,5,2,2,2,2,2,2,2,4
-4,2,2,5,3,3,3,2,2,2,2,4
-0,2,2,5,2,2,3,4,2,4,2,4
-1,1,1,2,1,1,3,2,1,1,3,4
-2,3,1,4,1,3,1,3,1,1,3,4
-4,3,1,6,4,3,2,5,1,1,3,4
-2,3,1,4,1,6,1,6,1,1,3,4
-3,1,1,2,3,1,1,5,2,1,3,4
-4,0,0,5,3,0,3,3,3,1,3,4
-3,2,1,6,3,3,1,6,1,2,3,4
-1,1,1,5,3,3,2,3,2,2,3,4
-4,4,4,5,1,1,1,2,5,5,3,4
-4,3,1,6,2,5,2,3,1,1,4,4
-5,3,1,5,1,2,1,5,1,1,4,4
-1,4,1,6,1,4,4,6,1,1,4,4
-4,2,1,5,3,2,2,4,4,1,4,4
-5,4,1,4,4,3,1,5,5,1,4,4
-3,3,2,4,3,0,3,4,0,4,4,4
-3,2,2,5,2,2,2,4,2,4,4,4
-4,3,2,5,1,3,2,6,4,4,4,4
-5,4,4,5,5,4,2,6,3,1,5,4
-4,4,2,5,5,5,2,6,4,1,5,4
-0,0,0,0,0,0,4,3,0,0,0,5
-0,0,0,0,0,4,0,4,0,0,0,5
-1,1,1,2,1,3,3,2,4,1,0,5
-0,0,0,2,0,0,0,3,0,4,0,5
-1,1,1,3,1,1,1,1,1,1,1,5
-1,1,1,4,1,1,2,1,1,1,1,5
-1,1,1,4,1,1,1,2,1,1,1,5
-1,1,1,4,1,1,1,2,1,1,1,5
-1,1,1,6,1,1,1,2,1,1,1,5
-1,1,1,2,1,3,3,2,1,1,1,5
-1,1,1,6,1,1,1,3,1,1,1,5
-1,1,1,3,1,2,1,3,1,1,1,5
-1,1,1,3,1,2,1,3,1,1,1,5
-3,1,1,5,2,2,1,3,1,1,1,5
-1,1,1,3,1,2,4,3,1,1,1,5
-1,1,1,6,1,4,4,3,1,1,1,5
-1,1,1,6,1,2,1,4,1,1,1,5
-2,2,1,4,2,2,2,4,1,1,1,5
-1,2,1,5,1,3,3,4,1,1,1,5
-2,2,2,4,1,2,1,5,1,1,1,5
-1,2,1,5,2,4,2,5,1,1,1,5
-4,3,1,5,4,4,1,6,1,1,1,5
-1,2,2,6,2,2,1,2,2,1,1,5
-0,1,1,3,2,1,3,2,2,1,1,5
-1,2,2,3,1,2,1,3,2,1,1,5
-1,1,1,4,1,1,2,3,2,1,1,5
-4,2,1,6,1,3,2,3,2,1,1,5
-1,1,2,3,1,2,1,4,2,1,1,5
-1,1,1,5,1,3,1,4,2,1,1,5
-1,2,1,5,2,3,1,4,2,1,1,5
-2,2,2,5,2,2,2,4,2,1,1,5
-1,2,1,6,2,2,1,5,2,1,1,5
-1,2,2,4,2,3,3,1,3,1,1,5
-2,2,1,5,2,3,2,4,3,1,1,5
-1,1,1,2,1,1,1,3,4,1,1,5
-1,2,2,5,1,2,2,4,4,1,1,5
-1,2,1,5,1,2,1,2,1,2,1,5
-1,1,1,3,3,3,2,2,1,2,1,5
-1,2,1,4,2,2,3,4,2,2,1,5
-1,2,1,4,3,3,3,4,2,2,1,5
-1,1,1,4,1,1,5,2,0,3,1,5
-2,1,1,2,1,3,4,1,1,3,1,5
-1,1,1,4,1,1,6,1,1,3,1,5
-1,1,1,4,1,2,2,2,1,3,1,5
-2,2,1,5,4,2,2,2,1,3,1,5
-2,1,1,3,1,1,3,2,1,3,1,5
-3,3,1,2,1,1,1,4,1,3,1,5
-1,1,1,3,1,2,4,5,4,3,1,5
-2,1,2,1,1,1,4,1,1,4,1,5
-1,1,1,2,1,1,1,3,1,4,1,5
-2,1,1,3,1,2,2,1,2,4,1,5
-1,1,2,5,1,2,3,3,3,4,1,5
-1,1,1,2,1,1,3,1,1,5,1,5
-1,1,1,5,1,1,5,1,1,5,1,5
-2,2,2,5,4,1,5,1,2,5,1,5
-1,1,1,0,1,1,1,1,1,6,1,5
-1,1,1,1,1,1,1,1,1,6,1,5
-1,1,1,5,1,1,5,1,1,6,1,5
-5,4,0,6,3,2,0,5,0,1,2,5
-4,4,1,6,1,3,4,2,1,1,2,5
-2,2,1,5,2,2,1,3,1,1,2,5
-3,2,2,3,1,1,4,3,1,1,2,5
-5,3,2,4,3,2,1,5,1,1,2,5
-1,3,1,3,4,2,1,5,1,1,2,5
-1,3,1,3,1,3,1,5,1,1,2,5
-1,2,1,6,1,3,1,5,1,1,2,5
-2,3,1,6,1,4,2,5,1,1,2,5
-2,4,1,5,2,4,2,5,1,1,2,5
-4,3,1,5,2,3,4,6,1,1,2,5
-1,1,1,4,1,3,2,3,2,1,2,5
-2,2,1,6,2,3,2,4,2,1,2,5
-4,1,2,6,4,4,4,1,3,1,2,5
-2,2,2,3,2,2,1,2,4,1,2,5
-1,2,1,5,2,1,2,4,1,2,2,5
-2,2,1,5,2,2,3,3,1,3,2,5
-0,3,1,5,1,1,3,4,1,3,2,5
-2,2,2,4,2,2,3,2,2,3,2,5
-3,1,1,4,1,1,1,1,3,3,2,5
-3,2,1,3,1,4,5,3,1,4,2,5
-3,2,1,3,1,2,1,4,1,4,2,5
-2,2,2,6,2,3,3,4,2,4,2,5
-4,2,1,3,2,3,5,2,3,4,2,5
-1,2,1,3,2,2,3,4,3,4,2,5
-3,3,3,3,3,3,4,3,4,4,2,5
-2,2,2,3,2,2,5,3,2,5,2,5
-2,2,2,5,2,5,5,3,2,5,2,5
-1,1,1,5,1,1,1,1,1,1,3,5
-4,3,2,5,2,3,1,5,1,1,3,5
-3,4,1,6,4,5,1,5,1,1,3,5
-4,4,1,6,5,5,1,5,1,1,3,5
-2,2,2,6,1,2,3,2,2,1,3,5
-1,2,1,5,1,3,4,3,2,1,3,5
-4,4,2,5,1,3,4,4,2,1,3,5
-4,2,1,5,1,2,1,5,2,1,3,5
-4,3,1,6,2,3,1,5,2,1,3,5
-3,3,1,6,3,4,1,5,2,1,3,5
-2,3,1,4,1,3,3,5,2,1,3,5
-4,3,3,6,2,3,4,3,3,1,3,5
-2,1,1,4,3,3,3,4,3,1,3,5
-3,1,3,4,3,3,4,4,3,3,3,5
-2,3,1,4,1,1,3,5,3,3,3,5
-1,1,1,2,3,1,4,1,1,4,3,5
-3,2,1,6,2,2,2,4,3,4,3,5
-4,3,2,5,3,4,4,3,5,4,3,5
-5,5,2,5,3,5,4,5,3,5,3,5
-2,5,1,6,1,3,1,2,4,5,3,5
-4,4,1,4,1,2,1,1,1,1,4,5
-2,3,1,6,2,3,1,4,1,1,4,5
-2,2,1,6,1,2,3,5,1,1,4,5
-3,4,1,5,2,3,3,5,1,1,4,5
-4,5,2,5,3,5,1,6,2,1,4,5
-5,4,1,5,3,5,1,4,3,1,4,5
-4,4,3,6,3,3,4,4,3,1,4,5
-3,5,3,4,4,4,4,4,3,5,4,5
-3,1,1,4,3,4,1,4,1,1,5,5
-5,5,1,6,2,5,3,4,1,1,5,5
-4,2,1,6,4,1,1,3,3,1,5,5
-3,1,1,6,1,4,5,1,1,5,5,5
-5,4,3,4,5,5,4,5,2,5,5,5
-6,2,2,4,4,4,5,4,6,5,6,5
-0,0,0,0,0,0,0,0,0,0,0,6
-0,0,0,0,0,0,0,0,0,0,0,6
-0,0,0,4,0,0,0,0,0,0,0,6
-4,1,3,0,0,5,0,2,0,0,0,6
-0,0,1,2,2,2,1,1,1,1,0,6
-4,3,2,3,1,2,2,2,2,1,0,6
-2,0,0,0,0,0,6,5,0,5,0,6
-4,3,2,5,3,3,5,3,3,6,0,6
-1,1,1,1,1,1,1,1,1,1,1,6
-1,1,1,2,1,1,1,1,1,1,1,6
-2,1,1,5,3,3,1,1,1,1,1,6
-0,1,1,5,2,2,2,1,1,1,1,6
-1,2,2,2,3,3,2,1,1,1,1,6
-1,1,1,1,1,1,4,1,1,1,1,6
-1,1,1,1,1,1,1,2,1,1,1,6
-1,1,1,4,1,1,1,2,1,1,1,6
-1,1,1,6,1,1,1,2,1,1,1,6
-1,1,1,4,1,2,2,2,1,1,1,6
-1,1,1,4,1,2,3,2,1,1,1,6
-1,1,2,3,2,2,3,2,1,1,1,6
-2,1,1,5,2,2,1,3,1,1,1,6
-1,1,1,2,1,2,3,3,1,1,1,6
-1,2,1,1,4,3,3,3,1,1,1,6
-1,1,1,5,1,2,1,4,1,1,1,6
-1,2,2,5,2,2,3,4,1,1,1,6
-1,1,1,5,3,3,5,4,1,1,1,6
-1,1,1,4,1,1,1,5,1,1,1,6
-1,1,1,5,1,1,6,5,1,1,1,6
-3,1,1,4,1,1,1,6,1,1,1,6
-1,1,1,4,1,2,1,4,2,1,1,6
-1,2,1,3,1,1,2,4,2,1,1,6
-1,2,1,5,1,2,4,4,2,1,1,6
-2,2,1,6,2,4,1,6,2,1,1,6
-1,1,1,4,1,1,2,2,3,1,1,6
-3,2,1,5,2,2,1,3,3,1,1,6
-1,2,1,5,1,3,5,4,3,1,1,6
-6,3,1,6,6,1,6,6,3,1,1,6
-3,2,3,5,1,3,3,3,4,1,1,6
-4,1,1,5,1,1,2,2,5,1,1,6
-1,1,1,2,1,1,2,1,1,2,1,6
-2,1,1,6,5,1,6,2,1,2,1,6
-2,2,2,6,5,2,4,1,2,2,1,6
-1,1,1,5,1,1,1,2,1,3,1,6
-1,1,1,4,1,1,3,4,1,3,1,6
-1,1,2,6,3,3,3,4,2,3,1,6
-1,2,1,5,3,3,5,3,4,3,1,6
-2,1,1,5,1,1,1,1,1,4,1,6
-1,1,1,2,1,1,4,1,1,4,1,6
-1,1,1,2,1,1,5,2,2,4,1,6
-1,2,1,4,1,3,0,4,2,4,1,6
-2,1,1,4,1,1,5,3,3,4,1,6
-1,1,1,4,2,2,6,4,3,4,1,6
-1,1,1,1,1,1,1,1,1,5,1,6
-1,1,1,3,1,1,3,1,1,5,1,6
-1,1,1,1,1,1,5,1,1,5,1,6
-1,1,1,3,2,2,5,1,2,5,1,6
-1,1,1,3,2,1,5,2,5,5,1,6
-1,1,1,6,4,4,6,1,1,6,1,6
-2,1,1,2,1,2,1,1,1,1,2,6
-2,1,1,6,5,5,2,1,1,1,2,6
-1,1,2,2,1,1,5,1,1,1,2,6
-2,3,1,4,1,1,1,3,1,1,2,6
-1,0,1,4,2,5,2,4,1,1,2,6
-1,3,1,5,1,3,2,5,1,1,2,6
-3,3,1,5,2,2,3,5,1,1,2,6
-0,2,2,3,1,4,3,4,2,1,2,6
-2,2,2,5,1,2,2,5,2,1,2,6
-3,3,3,4,1,3,1,4,3,1,2,6
-4,2,3,5,3,4,2,4,3,1,2,6
-0,1,1,5,1,2,2,1,4,1,2,6
-3,1,1,2,1,2,2,2,1,2,2,6
-1,2,1,5,1,2,3,4,2,2,2,6
-1,3,1,5,2,3,3,3,1,3,2,6
-1,1,2,3,1,2,3,2,2,4,2,6
-2,1,1,5,1,4,1,5,2,4,2,6
-1,2,3,4,1,3,5,2,1,5,2,6
-2,2,2,4,2,1,6,3,2,5,2,6
-5,1,1,2,1,1,6,2,5,5,2,6
-1,2,0,5,2,3,6,1,0,1,3,6
-3,2,2,5,2,2,2,4,2,1,3,6
-2,4,2,6,2,3,4,5,2,1,3,6
-2,3,3,5,1,3,5,2,2,2,3,6
-2,1,1,5,1,1,4,3,1,3,3,6
-2,2,1,5,2,2,2,5,1,3,3,6
-2,3,1,4,4,4,4,5,1,3,3,6
-3,2,1,5,3,3,2,6,1,3,3,6
-2,2,1,5,1,3,4,3,2,3,3,6
-4,2,1,5,2,3,3,1,1,4,3,6
-3,4,3,6,2,6,3,5,3,4,3,6
-3,3,3,4,3,3,4,5,3,4,3,6
-4,4,4,5,4,4,5,6,4,4,3,6
-4,3,2,5,2,3,4,3,1,6,3,6
-2,3,1,4,1,3,3,1,1,1,4,6
-4,5,2,5,1,4,4,5,1,1,4,6
-4,1,1,6,1,2,2,3,1,2,4,6
-4,2,1,5,4,3,5,3,2,4,4,6
-2,3,3,4,2,3,4,1,2,5,4,6
-1,1,2,5,3,4,1,2,2,1,5,6
-5,4,3,4,5,4,1,4,3,4,5,6
-5,3,4,2,3,4,5,5,4,5,5,6
-6,6,6,6,6,6,6,6,1,6,6,6
-    """.trimIndent()
-    val df = DataFrame.readCsvStr(csvString)
-
-
-
-    var selectedCategoryIndex by remember { mutableIntStateOf(0) }
-    val categories = listOf(
-        "François\nPiquemal",
-        "Jean-Luc\nMoudenc",
-        "François\nBriançon",
-        "Arthur\nCottrel",
-        "Julien\nLeonardelli",
-    )
-    val valuesPerCategory = listOf(
-        // FAC-SIMILE (FOR TESTING)
-        listOf(1.0f, -0.2f, 0.5f, -0.75f, -0.6f),
-        listOf(-0.2f, 1.0f, 0.5f, -0.75f, -0.6f),
-        listOf(0.35f, 0.0f, 1.0f, -0.75f, -0.6f),
-        listOf(-0.2f, 0.5f, -0.75f, 1.0f, -0.6f),
-        listOf(-0.2f, 0.5f, -0.75f, 0.6f, 1.0f),
-    )
-    JmTheme {
-        SpiderChart(
-            modifier = modifier
-                .height(400.dp)
-                .fillMaxWidth(),
-            title = {
-                PlotTitle(text = "Proximity of the proposals to " + categories[selectedCategoryIndex])
-            },
-            highlightedCategoryIndex = selectedCategoryIndex,
-            categories = categories,
-            values = valuesPerCategory[selectedCategoryIndex],
-            onCategoryClick = { clickedCategoryIndex ->
-                selectedCategoryIndex = clickedCategoryIndex
-            },
-        )
-        Text(df.columnNames().toString())
-    }
-}
